% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch13-fn.R
\name{anova1.diff}
\alias{anova1.diff}
\title{One-Way ANOVA: CI for the Difference in Means}
\usage{
anova1.diff(f, y, data, alp = 0.05, ci = "pair", dig = 4, ws = c(7, 4), main)
}
\arguments{
\item{f}{Object (aov), Formula (with data), or Factor (without data).}

\item{y}{Response variable (without data).}

\item{data}{Data frame applied to the formula.}

\item{alp}{Level of significance, Default: 0.05.}

\item{ci}{Method of CI, one of ('pair', 'hsd', 'both'), Default: 'pair'.}

\item{dig}{Number of digits below the decimal point, Default: 4.}

\item{ws}{Graphic window size for the box plot, Default: c(7,4).}

\item{main}{Title of the plot.}
}
\value{
None.
}
\description{
Estimating the confidence intervals for difference between means in one-way ANOVA.
}
\examples{
Yield <- c(79,83,88,78,75, 81,89,91,84,86,82, 86,91,93,90,89, 76,81,82,79)
Temp <- c(rep(150,5), rep(200,6), rep(250,5), rep(300,4))
# [1] Direct input
anova1.diff(f=Temp, y=Yield)
# [2] Formula input
anova1.diff(Yield ~ Temp)
# [3] anova1.fit object input
aa <- anova1.fit(f=Temp, y=Yield, ws="n", dp="n")
anova1.diff(aa)
# [4] aov (lm) object input
Temp <- as.factor(Temp)
dd <- aov(Yield ~ Temp)
anova1.diff(dd)
anova1.diff(dd, ci="hsd")
anova1.diff(dd, ci="both")
}
