% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch13-fn.R
\name{anova1.fit}
\alias{anova1.fit}
\title{One-Way ANOVA: Model Fitting and F-Test}
\usage{
anova1.fit(f, y, data, alp = 0.05, dig = 4, ws = c(7, 4), dp = 1:2, ...)
}
\arguments{
\item{f}{Formula (with data), or Factor (without data).}

\item{y}{Response variable (without data).}

\item{data}{Data frame applied to the formula.}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size for the box plot, Default: c(7,4).}

\item{dp}{Selection numbers of diagnostic plots, Default: 1:2.}

\item{...}{Other graphic parameters.}
}
\value{
Data and summary statistics.
}
\description{
One-way analysis of variance to produce an ANOVA table.
}
\examples{
Yield <- c(79,83,88,78,75, 81,89,91,84,86,82, 86,91,93,90,89, 76,81,82,79)
Temp <- c(rep(150,5), rep(200,6), rep(250,5), rep(300,4))
anova1.fit(f=Temp, y=Yield)
anova1.fit(Yield ~ Temp)
# 'mtcars' dataset
anova1.fit(mpg ~ cyl, data=mtcars)
}
