% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch13-fn.R
\name{anova2.conf}
\alias{anova2.conf}
\alias{anova2.diff}
\title{Two-Way ANOVA: CI for the Level Means}
\usage{
anova2.conf(f, data, alp = 0.05, maxim = TRUE, dig = 4, ws = c(7, 4), main)

anova2.diff(
  f,
  data,
  alp = 0.05,
  ci = "pair",
  sel = "h4",
  dig = 4,
  ws = c(7, 4),
  main
)
}
\arguments{
\item{f}{Object (aov), Formula (with data), or Object (anova2.fit).}

\item{data}{Data frame applied to the formula.}

\item{alp}{Level of significance, Default: 0.05.}

\item{maxim}{Logical: maximize the response? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size of the CI plot, Default: c(7,4).}

\item{main}{Title of the plot.}

\item{ci}{Method of CI, one of (pair, hsd, both), Default: 'pair'.}

\item{sel}{Selected combination levels, Default: 'h4' (high four).}
}
\value{
CI table.

CI table.
}
\description{
Estimating the confidence intervals for the level means in two-way ANOVA.

Estimating the confidence intervals for the difference in means in two-way ANOVA.
}
\examples{
Yield <- c(76,79,81,79,83,85, 79,81,84,86,89,88, 87,91,91,94,88,86, 79,82,85,84,77,76)
f1 <- c(rep(100,6), rep(150,6), rep(200,6), rep(250,6))
f2 <- rep(c(1,1,2,2,3,3), 4)
Temp <- as.factor(f1)
Press <- as.factor(paste0(f2, "Psig"))
anova2.conf(Yield ~ Temp * Press)

bb <- anova2.fit(Yield ~ Temp * Press)
anova2.conf(bb)

ee <- aov(Yield ~ Temp * Press)
anova2.conf(ee)

# Without replication
Yield1 <- c(77.5,80.0,89.0,80.5, 80.0,85.0,92.5,84.5, 84.0,88.5,87.0,76.5)
Temp1 <- c(100,150,200,250, 100,150,200,250, 100,150,200,250)
Press1 <- c(1,1,1,1, 2,2,2,2 ,3,3,3,3)
anova2.conf(Yield1 ~ Temp1 * Press1)

cc <- anova2.fit(Yield1 ~ Temp1 * Press1)
anova2.conf(cc)

Temp.f <- as.factor(Temp1)
Press.f <- as.factor(Press1)
ff <- aov(Yield1 ~ Temp.f + Press.f)
anova2.conf(ff)
Yield <- c(76,79,81,79,83,85, 79,81,84,86,89,88, 87,91,91,94,88,86, 79,82,85,84,77,76)
f1 <- c(rep(100,6), rep(150,6), rep(200,6), rep(250,6))
f2 <- rep(c(1,1,2,2,3,3), 4)
Temp <- as.factor(f1)
Press <- as.factor(paste0(f2, "Psig"))
anova2.diff(Yield ~ Temp * Press)

bb <- anova2.fit(Yield ~ Temp * Press)
anova2.diff(bb)

ee <- aov(Yield ~ Temp * Press)
anova2.diff(ee)

# Without replication
Yield1 <- c(77.5,80.0,89.0,80.5, 80.0,85.0,92.5,84.5, 84.0,88.5,87.0,76.5)
Temp1 <- c(100,150,200,250, 100,150,200,250, 100,150,200,250)
Press1 <- c(1,1,1,1, 2,2,2,2 ,3,3,3,3)
anova2.diff(Yield1 ~ Temp1 * Press1)

cc <- anova2.fit(Yield1 ~ Temp1 * Press1)
anova2.diff(cc)

Temp.f <- as.factor(Temp1)
Press.f <- as.factor(Press1)
ff <- aov(Yield1 ~ Temp.f + Press.f)
anova2.diff(ff)
}
