% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch13-fn.R
\name{anova2.fit}
\alias{anova2.fit}
\title{Two-Way ANOVA: Model Fitting and F-Test}
\usage{
anova2.fit(form, data, alp = 0.05, dig = 4, ws = c(7, 4), dp = 1:2, ...)
}
\arguments{
\item{form}{Formula of the two-way ANOVA.}

\item{data}{Data frame applied to the formula.}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size for the plots, Default: c(7,4).}

\item{dp}{Selection numbers of diagnostic plots, Default: 1:2.}

\item{...}{Other graphic parameters.}
}
\value{
List of fitted results.
}
\description{
Two-way analysis of variance to produce an ANOVA table.
}
\examples{
Yield <- c(76,79,81,79,83,85, 79,81,84,86,89,88, 87,91,91,94,88,86, 79,82,85,84,77,76)
f1 <- c(rep(100,6), rep(150,6), rep(200,6), rep(250,6))
f2 <- rep(c(1,1,2,2,3,3), 4)
Temp <- as.factor(f1)
Press <- as.factor(paste0(f2, "P"))
anova2.fit(Yield ~ Temp * Press)

# Without replication
Yield1 <- c(77.5,80.0,89.0,80.5, 80.0,85.0,92.5,84.5, 84.0,88.5,87.0,76.5)
Temp1 <- c(100,150,200,250, 100,150,200,250, 100,150,200,250)
Press1 <- c(1,1,1,1, 2,2,2,2 ,3,3,3,3)
anova2.fit(Yield1 ~ Temp1 + Press1)
}
