% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{prop1.test}
\alias{prop1.test}
\title{Hypothesis Test for a Proportion}
\usage{
prop1.test(
  x,
  n,
  p0,
  alt = "two",
  method = "binom",
  alp = 0.05,
  dig = 4,
  ws = c(7, 4),
  ...
)
}
\arguments{
\item{x}{Number of successes.}

\item{n}{Sample size.}

\item{p0}{Population proportion under H0 (if missing, just find the CI).}

\item{alt}{Type of the alternative hypothesis ("gr", "le", "two"), Default: "two".}

\item{method}{One of ("binom", "all"), Default: "binom".}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
Summary Statistics.
}
\description{
Hypothesis Test for a Proportion by Exact and Approximation Methods
}
\examples{
# Confidence Intervals
prop1.test(x=15, n=200)
prop1.test(x=15, n=200, method="all")
prop1.test(x=11:15, n=200, method="all")
prop1.test(x=15, n=200, alp=c(0.1, 0.05, 0.01), method="all")

# Investigate the p-values
prop1.test(x=2:4, n=10, p0=0.1, alt="gr")
prop1.test(x=6:4, n=20, p0=0.5, alt="two")

# Check the accuracy of normal approximation 
prop1.test(x=15, n=200, p0=0.1, alt="le", method="all")
prop1.test(x=15, n=200, p0=0.1,  alt="two", method="all")

}
