% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{chi.prob}
\alias{chi.prob}
\title{Cumulative Probability of the Chi-square Distribution}
\usage{
chi.prob(nu, ub, lb = 0, prt = TRUE, dig = 4, ws = c(7, 4), ...)
}
\arguments{
\item{nu}{Degree of freedom of the chi-square distribution.}

\item{ub}{Values of upper limits for F(ub).}

\item{lb}{Values of lower limits for P(lb<X<ub), Default: 0.}

\item{prt}{Logical: print the output? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Grapic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
Cumulative (or Interval) Probabilities
}
\description{
Cumulative Probability of the Chi-square Distribution.
}
\details{
DETAILS
}
\examples{
k <- 1:10; nu <- 5
chi.prob(nu, ub=k)

nu <- 5; k <- seq(1, 15, by=2)
chi.prob(nu, ub=k)
}
