% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{chi.quant}
\alias{chi.quant}
\title{Quantile Plot of the Chi-square Distribution}
\usage{
chi.quant(nu, pv, prt = TRUE, dig = 4, ws = c(7, 4), PDF = TRUE, ...)
}
\arguments{
\item{nu}{Degree of freedom of the chi-square distribution}

\item{pv}{Probability values to find the quantiles.}

\item{prt}{Logical: print the output? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Grapic window size, Default: c(7,4).}

\item{PDF}{Logical: plot the PDF? (o.w. CDF) Default: TRUE.}

\item{...}{Other graphic parameters.}
}
\value{
Quantiles of the Chi-square Distribution.
}
\description{
Quantile Plot of the Chi-square Distribution
}
\examples{
nu <- 5
pv <- c(0.005, 0.01, 0.025, 0.05, 0.5, 0.95, 0.975, 0.99, 0.995)
chi.quant(nu, pv)
chi.quant(nu, pv, PDF=FALSE)
}
