% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{chi.sim}
\alias{chi.sim}
\title{Simulation for the chi-square Distribution}
\usage{
chi.sim(
  ns,
  mu = 0,
  sig = 1,
  N = 1000,
  muknow = TRUE,
  prt = TRUE,
  ws = c(10, 4),
  dig = 4,
  ...
)
}
\arguments{
\item{ns}{Sample size.}

\item{mu}{Expected value, Default: 0.}

\item{sig}{Standard deviation, Default: 1.}

\item{N}{Number of iterations, Default: 1000.}

\item{muknow}{Logical: is mu known?, Default: TRUE.}

\item{prt}{Logical: print statistics? Default: TRUE.}

\item{ws}{Graphic window size, Default: c(10,4).}

\item{dig}{Number of decimal places, Default: 4.}

\item{...}{Other graphic parameters.}
}
\value{
Sample means and standard deviations.
}
\description{
Simulation for the chi-square Distribution
}
\examples{
chi.sim(ns=5)
chi.sim(ns=5, muknow=FALSE)

chi.sim(ns=10, mu=100, sig=10)
chi.sim(ns=10, mu=100, sig=10, muknow=FALSE)

chi.sim(ns=5, N=10000, dig=4, breaks=100, seed=1243)
chi.sim(ns=5, N=10000, dig=4, muknow=FALSE, breaks=100, seed=1243)

}
