% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{chigof.test}
\alias{chigof.test}
\title{Goodness of Fit Test from a Frequency Table}
\usage{
chigof.test(x, p, alp = 0.05, dtab = 2, dig = 4, ws = c(7, 4))
}
\arguments{
\item{x}{Frequency table.}

\item{p}{Probability vector, Default: rep(1/k, k).}

\item{alp}{Level of significance, Default: 0.05.}

\item{dtab}{Number of decimal places for the table, Default: 2.}

\item{dig}{Number of decimal places for other output, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}
}
\value{
list(stat, df, cv, pv, tab)
}
\description{
Goodness of fit test from a table, when the cell probabilities are given.
}
\examples{
# Goodness-of-fit Test
x <- c(31,26,22,18,13,10)
chigof.test(x)

x <- c(32, 65, 47, 38, 18)
p <- c(0.15, 0.3, 0.25, 0.2, 0.1)
chigof.test(x, p)

}
