% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{chitest.plot2}
\alias{chitest.plot2}
\title{Plot the Chi-square Test}
\usage{
chitest.plot2(stat, df, alp = 0.05, alt = "gr", dig = 4, ...)
}
\arguments{
\item{stat}{Chi-square test statistic.}

\item{df}{Degrees of freedom.}

\item{alp}{Level of significance, Default: 0.05.}

\item{alt}{Type of the alternative hypothesis, Default: 'gr'.}

\item{dig}{Number of decimal places, Default: 4.}

\item{...}{Other graphic parameters.}
}
\value{
None.
}
\description{
Plot the result of a chi-square test.
}
\examples{
# Goodness-of-fit Test
x <- c(31,26,22,18,13,10)
(ct <- chisq.test(x))
chitest.plot2(stat=ct$stat, df=ct$para)

# Test of Homogeneity
x <- c(20,16,29,21,14,  14,22,26,25,13,  18,24,32,18, 8,  8,18,33,16,25)
x <- matrix(x, nrow=4, ncol=5, byrow=TRUE)
(ct <- chisq.test(x))
chitest.plot2(stat=ct$stat, df=ct$para)

}
