% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{ciprop.sim}
\alias{ciprop.sim}
\title{Simulate the Confidence Interval for a Proportion}
\usage{
ciprop.sim(
  n,
  p = 0.5,
  alp = 0.05,
  N = 100,
  seed = 9857,
  ws = c(7, 4),
  method = "norm"
)
}
\arguments{
\item{n}{Sample size.}

\item{p}{Population proportion value, Default: 0.5.}

\item{alp}{Level of significance, Default: 0.05.}

\item{N}{Number of iterations, Default: 100.}

\item{seed}{Seed value for generating random numbers, Default: 9857.}

\item{ws}{Graphic window size, Default: c(7,4).}

\item{method}{One of ("normal", "wilson", "binom"), Default: "normal".}
}
\value{
Summary Statistics.
}
\description{
Simulate the Confidence Interval for a Population Proportion
}
\examples{
ciprop.sim(n=16, p=0.6, alp=0.05, N=100)
ciprop.sim(n=16, p=0.6, alp=0.05, N=10000, ws="n")

ciprop.sim(n=100, p=0.6, alp=0.05, N=10000, ws="n")
}
