% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{civar2.sim}
\alias{civar2.sim}
\title{Simulate the Confidence Interval for a the Ratio of Variances}
\usage{
civar2.sim(
  var1,
  var2,
  n1,
  n2,
  alp = 0.05,
  N = 100,
  seed = 9857,
  dig = 4,
  ws = c(7, 4)
)
}
\arguments{
\item{var1}{Variance of the first population.}

\item{var2}{Variance of the second population.}

\item{n1}{Sample size of the first population.}

\item{n2}{Sample size of the second population.}

\item{alp}{Level of significance, Default: 0.05.}

\item{N}{Number of iterations, Default: 100.}

\item{seed}{Seed value for generating random numbers, Default: 9857.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}
}
\value{
None.
}
\description{
Simulate the Confidence Interval for a the Ratio of Population Variances
}
\examples{
civar2.sim(n1=25, n2=16)
civar2.sim(var1=8, var2=2, n1=25, n2=16)
civar2.sim(var1=8, var2=2, n1=25, n2=16, N=10000, plot=F)
}
