% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{clt.plot}
\alias{clt.plot}
\title{Diagnosis of the Central Limit Theorem}
\usage{
clt.plot(dist, para, para2, ns, N = 1000, sigknow = TRUE, ...)
}
\arguments{
\item{dist}{Name of population distribution ("exp","gamma","weibull","beta","norm", "t","chisq","f","pois","binom")}

\item{para}{First parameter of the distribution.}

\item{para2}{Second parameter of the distribution. (if necessary)}

\item{ns}{Sample size, Default: c(10, 30, 50).}

\item{N}{Number of iterations, Default: 10000.}

\item{sigknow}{Logica: is population variance known? Default: TRUE.}

\item{...}{Other graphic parameters. (seed, col, fill, lwd, pch, pt.col)}

\item{d}{Group width in histogram, Default: rep(0.5, 3).}

\item{seed}{Seed value for generating random numbers, Default: 9857}
}
\value{
None.
}
\description{
Diagnosis of the Central Limit Theorem
}
\examples{
# t-distribution
clt.plot("t", para=3)
clt.plot("t", para=3, breaks=50)
clt.plot("t", para=3, N=10000, breaks=50)
# Exponential distribution
clt.plot("exp", para=5, N=10000)
clt.plot("exp", para=5, ns=c(1,3,5,10)*10, breaks=50)
# Gamma distribution
clt.plot("gam", para=0.5, para2=10)
clt.plot("gam", para=0.5, para2=10, ns=c(1,3,5,10)*10, breaks=50)
clt.plot("gam", para=0.5, para2=10, ns=c(1,3,5,10)*10, N=10000, breaks=100)
}
