% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{condprob}
\alias{condprob}
\title{Conditional probabilities from a Two-way table}
\usage{
condprob(tab, evt, cev, prt = TRUE, debug = FALSE)
}
\arguments{
\item{tab}{Two-way frequency table with unique col-names and raw-names.}

\item{evt}{Vector of event names.}

\item{cev}{Vector of conditional event names.}

\item{prt}{Logical value for detailed output, Default: TRUE.}
}
\description{
Obtain conditional probabilities from a Two-way table.
}
\examples{
# Create a two-way frequency table
set.seed(123)
tab = as.table(matrix(sample(10:30, 12), ncol=4))
colnames(tab) <- c("V", "W", "X", "Y")
rownames(tab) <- c("A", "B", "C")
addmargins(tab)
# P(A|X), P(X|A)
condprob(tab, "A", "X")
condprob(tab, "X", "A")
# P(A or B|X or Y)
condprob(tab, c("A","B"), c("X","Y"))
}
