% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.cdf}
\alias{cont.cdf}
\alias{cont.prob}
\title{CDF of a Continuous Random Variable}
\usage{
cont.cdf(fun, lb, ub, xp, prt = FALSE, dig = 4, ws = c(7, 5), xlim, ...)

cont.prob(fun, aa, bb, lb, ub, prt = TRUE, dig = 4, ws = "n", xlim, ...)
}
\arguments{
\item{fun}{Probability density function f(x) or "f(x)".}

\item{lb}{Lower bound of f(x)}

\item{ub}{Upper bound of f(x)}

\item{xp}{Specific values of X for displaying the probability.}

\item{prt}{Logical: print F(xp)? Default: FALSE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,5).}

\item{xlim}{Lower and Upper limit of x-axis.}

\item{...}{Other graphic parameters.}

\item{aa}{Upper bound bb of P(aa<X<bb).}

\item{fx}{Probability density function f(x) of X.}
}
\value{
list(CDF=CDF, mean=EX, sd=DX).

Table of P(X<bb), P(X<aa), and P(aa<X<bb).
}
\description{
Cumulative Distribution Function of a Continuous Random Variable.

Interval Probability P(aa<X<bb) of a Continuous Random Variable X.
}
\examples{
# f(x) = 2*exp(-2*x), for (x>0)
cont.cdf("2*exp(-2*x)", lb=0, xp=c((1:5)*0.2, 2), xlim=c(-1,3))
cont.cdf("2*exp(-2*x)", lb=0, xp=c((1:5)*0.2, 2), prt=TRUE, ws="n")
cont.cdf("2*exp(-2*x)", lb=0, xp=c((1:5)*0.2, 2), prt=TRUE)

# f(x) = 3*x^2, for (0<x<1)
cont.cdf("3*x^2", lb=0, ub=1, xp=(2:5)*0.2, prt=TRUE)

# f(x) = dnorm(x,10,2), for (-Inf<x<Inf)
cont.cdf("dnorm(x,10,2)", xp=3:7*2, prt=TRUE)

pdf <- function(x) 2*exp(-2*x)*(x>0)
cont.prob(pdf, aa=c(0.2,0.7), bb=c(0.5,1))
cont.prob(pdf, aa=c(0.2,0.7), bb=c(0.5,1), prt=FALSE, ws=c(7,4))
cont.prob(pdf, aa=c(0.2,0.7), bb=c(0.5,1), ws=c(7,4), xlim=c(-0.5,2))

# f(x) = 3*x^2, for (0<x<1)
cont.prob("3*x^2", aa=0:2*0.3, bb=1:3*0.25, lb=0, ub=1, ws=c(7,4))

# Other built-in PDFs
cont.prob("dnorm(x,10,2)", aa=1:3*4, bb=1:3*4+3, ws=c(7,4))
cont.prob("dexp(x)", aa=1:3*0.4, bb=1:3*0.4+0.3, prt=TRUE, ws=c(7,4))

}
