% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.dint}
\alias{cont.dint}
\title{Double Integration of Joint PDF of Continuous Random Variables}
\usage{
cont.dint(fun, x1, x2, y1, y2, frac = FALSE, Big = 100)
}
\arguments{
\item{fun}{Continuous joint PDF as string "f(x,y)".}

\item{x1}{Lower limit (numeric) of X.}

\item{x2}{Upper limit (numeric) of X.}

\item{y1}{Lower limit (numeric, function or strings) of Y.}

\item{y2}{Upper limit (numeric, function or strings) of Y.}

\item{frac}{Logical: convert the result to fraction? Default: FLASE.}

\item{Big}{Big number replacing Inf (functional bound case), Default: 100.}
}
\value{
Double integral.
}
\description{
Double Integration of Joint PDF of Continuous Random Variables.
}
\examples{
# Integrate f(x,y) = 2*(x+y)*(x<y) over 0<x,y<1
cont.dint("2*(x+y)", 0, 1, "x", 1)
# P(0<X<0.5, 0<Y<0.5)
cont.dint("2*(x+y)", 0, 0.5, "x", 0.5)
cont.dint("2*(x+y)", 0, 0.5, "x", 0.5, frac=TRUE)

# Integrate f(x,y) = 2*exp(-x-y) over 0<x<y
cont.dint("2*exp(-x-y)", 0, Inf, "x", Inf)
# P(0<X<0.5, X<Y<0.5)
cont.dint("2*exp(-x-y)", 0, 0.5, "x", 0.5)

}
