% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch7-fn.R
\name{cont.dist}
\alias{cont.dist}
\title{Expectations, Transformation, and PDF Plots for Continuous Distributions}
\usage{
cont.dist(dist, para1, para2, Tr, prt = "var", dig = 4, ws = "n", ...)
}
\arguments{
\item{dist}{Name of continuous probability distribution (one of the follows)
("unif", "exp", "gamma", "weibull", "beta", "norm", "t", "chisq", "f").}

\item{para1}{First parameter vector of the distribution.}

\item{para2}{Second parameter vector of the distribution (optional).}

\item{Tr}{Transformation function in text string.}

\item{prt}{Print option, one of c("exp", "var", "n"), Default: "var".}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graph window size, Default: "n".}

\item{...}{Other graphic parameters.}
}
\value{
list(fun, Ex, Vx, Tfun, Ey, Vy)
}
\description{
Expectation, Variance, Transformation, and PDF Plots for Continuous Distributions.
}
\examples{
# X ~ U(0,1), Y = X^2
cont.dist("unif", 0, 1, Tr="x^2", prt="var", ws=c(9,4))
# X ~ Exp(5), Y = 2X+4
cont.dist("exp", 5, Tr="2*x+4", prt="var", ws=c(9,4))
# X ~ Gamma(2,5), Y = 2(X-5)
cont.dist("gamma", 2, 5, Tr="2*(x-5)", prt="var", ws=c(9,4))
# X ~ Weibull(2,5), Y = X^2
cont.dist("weib", 2, 5, Tr="x^2", prt="var", ws=c(9,4))
# X ~ Beta(2,5), Y = 2+4X
cont.dist("beta", 2, 5, Tr="2+4*x", prt="var", ws=c(9,4))

}
