% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{cont.exp}
\alias{cont.exp}
\title{Expected Values of a Continuous Random Variable}
\usage{
cont.exp(fun, lb, ub, Tr, prt = "exp", dig = 4, ws = "n", ...)
}
\arguments{
\item{fun}{Continuous PDF (function or string).}

\item{lb}{Lower bound of fun, Default: -Inf.}

\item{ub}{Upper bound of fun, Default: Inf.}

\item{Tr}{Transformation formula in string.}

\item{prt}{Print option: one of c("exp", "var"), Default: "exp".}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}

\item{...}{Other graphic parameters.}
}
\value{
list(EX=E(X), VX=Var(X))
}
\description{
Calculate E(X) and Var(X) of a Continuous Random Variable X
}
\examples{
pdf <- function(x) exp(-2*x)*(x>0)
cont.exp(pdf, prt="var")

cont.exp("2*exp(-2*x)", lb=0, prt="var")

cont.exp("2*y", lb=0, ub=1, prt="var")

cont.exp("dnorm(x, 10, 3)", prt="var")
cont.exp(\(x) dnorm(x, 10, 3), prt="var")

y="x^2"; cont.exp("dnorm(x)", Tr=y, prt="var")

}
