% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.jcdf}
\alias{cont.jcdf}
\title{Joint CDF of Two Continuous Random Variables}
\usage{
cont.jcdf(fun, x1, y1, y2, xp, yp, ws = "n", xrng, yrng, dig = 4)
}
\arguments{
\item{fun}{Continuous joint PDF as string "f(x,y)".}

\item{x1}{Lower limit (numeric) of X.}

\item{y1}{Lower limit (numeric or character) of Y.}

\item{y2}{Upper limit (character) of Y.}

\item{xp}{Values of X for obtaining F(xp,yp)}

\item{yp}{Values of Y for obtaining F(xp,yp)}

\item{ws}{Graphic window size, Default: "n".}

\item{xrng}{Plot range of x-axis}

\item{yrng}{Plot range of y-axis}

\item{dig}{Number of decimal places, Default: 4.}
}
\value{
data frame(x=xp, y=yp, F=F(xp,yp)).
}
\description{
Joint CDF F(xp,yp) of Two Continuous Random Variables.
}
\examples{
# CDF for f(x,y)=2*(x+y), (0<x<y<1)
cont.jcdf("2*(x+y)", 0, "x", xp=-1:6/5, yp=-1:6/5)
# Plot the CDF in 3D
cont.jcdf("2*(x+y)", 0, "x", ws=c(7,5), xrng=c(0,1), yrng=c(0,1))

}
