% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{cont.jexp}
\alias{cont.jexp}
\title{Expected Values of Two Continuous Random Variables}
\usage{
cont.jexp(fun, lb, ub, y1, y2, Tr, prt = "exp", dig = 4, frac = FALSE)
}
\arguments{
\item{fun}{Continuous joint PDF as string "f(x,y)".}

\item{lb}{Lower bound (numeric) of X and Y.}

\item{ub}{Upper bound (numeric) of X and Y.}

\item{y1}{Lower limit (function or strings) of Y.}

\item{y2}{Upper limit (function or strings) of Y.}

\item{Tr}{Transformation formula in string.}

\item{prt}{Print option: one of c("exp", "var", "cov", "cor"), Default: "exp".}

\item{dig}{Number of decimal places, Default: 4.}

\item{frac}{Logical: convert the result to fraction? Default: FALSE.}
}
\value{
list(EX=E(X), EY=E(Y), VX=V(X), VY=V(Y), EXY=E(X,Y), VXY=Cov(X,Y), CXY=Corr(X,Y))
}
\description{
Joint pdf and Expected Values of Two Continuous Random Variables
}
\examples{
# [Ex1] f(x,y) = 0.5*(x+3*y), (0<x<1, 0<y<1)
cont.jexp("(x+3*y)", 0, 1, prt="cor")

# [Ex2] f(x,y) = 2*(x+y), (0<x<y<1)
cont.jexp("(x+y)", 0, 1, y1="x", prt="cor")

# [Ex3] fxy = 2*exp(-(x+y)), (0<x<y<Inf)
cont.jexp("exp(-(x+y))", 0, Inf, y1="x", prt="cor")

# [Ex4] fxy = 24*x*y, (0<x<1, 0<y<1, x+y<1)
cont.jexp("x*y", 0, 1, y2="1-x", prt="cor")

}
