% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.jpdf}
\alias{cont.jpdf}
\title{Joint PDF of Two Continuous Random Variables}
\usage{
cont.jpdf(
  fun,
  lb,
  ub,
  y1,
  y2,
  type = "m",
  xp,
  yp,
  prt = TRUE,
  dig = 4,
  ws = "n",
  xl,
  yl
)
}
\arguments{
\item{fun}{Continuous joint PDF function "f(x,y)".}

\item{lb}{Lower bounds of X and Y, numeric.}

\item{ub}{Upper bounds of X and Y, numeric.}

\item{y1}{Functional lower limit of Y, as "f(x)".}

\item{y2}{Functional upper limit of Y, as "f(y)".}

\item{type}{"m"=marginal, "c"=conditional, "i"=independence, Default: "m".}

\item{xp}{specific (or conditional) value of X.}

\item{yp}{specific (or conditional) value of Y.}

\item{prt}{Logical: print the marginal PDF? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}

\item{xl}{Limit of x-axis of the plot.}

\item{yl}{Limit of y-axis of the plot.}
}
\value{
Marginal("m") or conditional("c") PDF as functions.
}
\description{
Obtain Marginal and Conditional PDF and Check Independence from the Joint PDF of Two Continuous Random Variables as Functions.
}
\examples{
# [Ex1] f(x,y) = (x+y), (0<x<1, 0<y<1)
# Marginal PDF
fn <- cont.jpdf("x+y", 0, 1, type="m")
integrate(fn$fx, 0, Inf)[[1]]; integrate(fn$fy, 0, Inf)[[1]]
fn <- cont.jpdf("x+y", 0, 1, type="m", prt=FALSE, 
                ws=c(7,5), xp=1:4/5, yp=1:4/5)
# Conditional PDF
fn <- cont.jpdf("x+y", 0, 1, type="c")
fn <- cont.jpdf("x+y", 0, 1, xp=1:4/5, type="c")
fn <- cont.jpdf("x+y", 0, 1, xp=1:4/5, type="c", prt=FALSE, ws=c(7,5))
# Independence
fn=cont.jpdf("x+y", 0, 1, type="i")

# [Ex2] f(x,y) = 2*(x+y), (0<x<y<1)
# Marginal PDF
fn <- cont.jpdf("2*(x+y)", 0, 1, y1="x", type="m")
integrate(fn$fx, 0, Inf)[[1]]; integrate(fn$fy, 0, Inf)[[1]]
fn <- cont.jpdf("2*(x+y)", 0, 1, y1="x", type="m", prt=FALSE,
                ws=c(7,5),xp=0.2*2:4,yp=0.2*2:4)
# Conditional PDF
fn <- cont.jpdf("2*(x+y)", 0, 1, y1="x", type="c")
fn <- cont.jpdf("2*(x+y)", 0, 1, y1="x", xp=1:4/5, type="c")
fn <- cont.jpdf("2*(x+y)", 0, 1, y1="x", xp=1:4/5, type="c", 
                prt=FALSE, ws=c(7,5))
# Independence
fn <- cont.jpdf("2*(x+y)", 0, 1, y1="x", type="i")
# [Ex3] fxy = 2*exp(-(x+y)), (0<x<y<Inf)
# Marginal PDF
fn <- cont.jpdf("2*exp(-(x+y))", 0, Inf, y1="x")
integrate(fn$fx, 0, Inf)[[1]]; integrate(fn$fy, 0, Inf)[[1]]
# Conditional PDF
fn <- cont.jpdf("2*exp(-(x+y))", 0, Inf, y1="x", type="c")
fn <- cont.jpdf("2*exp(-(x+y))", 0, Inf, y2="x", xp=1:4, type="c", 
                prt=FALSE, ws=c(7,5), xl=c(0,6))
# Independence
fn <- cont.jpdf("2*exp(-(x+y))", 0, Inf, y1="x", type="i")

# [Ex4] fxy = 24*x*y, (0<x<1, 0<y<1, x+y<1)
# Marginal PDF
fn <- cont.jpdf("24*x*y", 0, 1, y2="1-x")
integrate(fn$fx, 0, Inf)[[1]]; integrate(fn$fy, 0, Inf)[[1]]
# Conditional PDF
fn <- cont.jpdf("24*x*y", 0, 1, y2="1-x", type="c")
fn <- cont.jpdf("24*x*y", 0, 1, y2="1-x", type="c", 
                prt=FALSE, ws=c(7,5), yp=1:4/5)
# Independence
fn <- cont.jpdf("24*x*y", 0, 1, y2="1-x", type="i")
}
