% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch7-fn.R
\name{cont.mpdf}
\alias{cont.mpdf}
\title{PDF and CDF Plots for Continuous Distributions}
\usage{
cont.mpdf(dist, para1, para2, xlim, ymax, xp1, xp2, np = 100, ws, ...)
}
\arguments{
\item{dist}{Name of continuous probability distribution (one of the follows)
("exp", "gamma", "weibull", "beta", "norm", "t", "chisq", "f").}

\item{para1}{First parameter vector of the distribution.}

\item{para2}{Second parameter vector of the distribution (optional).}

\item{xlim}{Limit of x-axis for the plot.}

\item{ymax}{Upper limit of y-axis for the plot.}

\item{xp1}{Specific x values for annotations in the PDF plot.}

\item{xp2}{Specific x values for annotations in the CDF plot.}

\item{np}{Number of plot points, Default: 100.}

\item{ws}{Graph window size, Default: c(9,5).}

\item{...}{Other graphic parameters.}
}
\value{
None.
}
\description{
PDF and CDF Plots for Continuous Distributions.
}
\examples{
# Five exponential PDFs
lamb <- 1:5
cont.mpdf("exp", para1=lamb, xlim=3, ymax=5)
# Four gamma PDFs
alp <- c(0.5, 1, 2, 3); rate <- 1
cont.mpdf("gamma", alp, rate, xlim=8, ymax=1.2)
# Four Weibull PDFs
th <- 1; alp <- c(0.5, 1, 2, 3)
cont.mpdf("weibull", alp, th, 5, 1.2)
# PDFs with rainbow colors
alp <- 10:100/10; th <- 1
cont.mpdf("gamma", alp, th, 20)
cont.mpdf("weibull", alp, th, 5)
}
