% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch7-fn.R
\name{cont.plot}
\alias{cont.plot}
\title{PDF Plot of Continuous Random Variables}
\usage{
cont.plot(FUN, lo, up, mt, dig = 3, xn = "X", prt = FALSE, pos = "center")
}
\arguments{
\item{FUN}{Continuous probability density function}

\item{lo}{Lower limit of x-axis}

\item{up}{Upper limit of x-axis}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 3}

\item{xn}{Random variable name, Default: 'X'}

\item{prt}{Print expected values and variances? Default: FALSE}

\item{pos}{Legend location, Default: 'center'}
}
\value{
list(ev=Ex, std=Dx)
}
\description{
PDF Plot of Continuous Random Variables
}
\examples{
fx <- function(x) dnorm(x,10,2)
cont.plot(fx, 0, 20, prt=T)

fx <- function(x) dgamma(x, 3, scale=5)
cont.plot(fx, 0, 40, mt="X ~ Gam(3,5)", prt=T)
}
