% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.prob}
\alias{cont.prob}
\title{Interval Probability of a Continuous Random Variable}
\usage{
cont.prob(fun, aa, bb, lb, ub, prt = TRUE, dig = 4, ws = "n", xlim, ...)
}
\arguments{
\item{fun}{Probability density function f(x) or "f(x)".}

\item{aa}{Upper bound bb of P(aa<X<bb).}

\item{lb}{Lower bound of f(x)}

\item{ub}{Upper bound of f(x)}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,5).}

\item{xlim}{Lower and Upper limit of x-axis.}

\item{...}{Other graphic parameters.}
}
\value{
Table of P(X<bb), P(X<aa), and P(aa<X<bb).
}
\description{
Interval Probability P(aa<X<bb) of a Continuous Random Variable X.
}
\examples{
# f(x) = 2*exp(-2*x) for (x>0)
pdf <- function(x) 2*exp(-2*x)*(x>0)
cont.prob(pdf, aa=c(0.2,0.7), bb=c(0.5,1))
cont.prob(pdf, aa=c(0.2,0.7), bb=c(0.5,1), prt=FALSE, ws=c(7,4))
cont.prob(pdf, aa=c(0.2,0.7), bb=c(0.5,1), ws=c(7,4), xlim=c(-0.5,2))

# f(x) = 3*x^2, for (0<x<1)
cont.prob("3*x^2", aa=0:2*0.3, bb=1:3*0.25, lb=0, ub=1, ws=c(7,4))

# Other built-in PDFs
cont.prob("dnorm(x,10,2)", aa=1:3*4, bb=1:3*4+3, ws=c(7,4))
cont.prob("dexp(x)", aa=1:3*0.4, bb=1:3*0.4+0.3, prt=TRUE, ws=c(7,4))

}
