% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.trans}
\alias{cont.trans}
\title{Transformed PDF of a Continuous Random Variable}
\usage{
cont.trans(fx, flim, tf, a, b, prt = TRUE, dig = 4, ws = "n", xlim, ...)
}
\arguments{
\item{fx}{PDF of X as string "f(x)".}

\item{flim}{range of X for f(x) > 0.}

\item{tf}{Character vector of transformation functions (up to 8).}

\item{a}{Lower limit for calculating P(a<X<b).}

\item{b}{Upper limit for calculating P(a<X<b).}

\item{prt}{Logical: print the output? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Window size of the plot, Default: "n".}

\item{xlim}{range of X for the plot.}

\item{...}{Other graphic parameters for text.}
}
\value{
Transformed PDFs.
}
\description{
Transformed PDF of a Continuous Random Variable.
}
\examples{
# Linear Transformation
TR <- c(y="10*x-4", w="-10*x+4")
cont.trans(fx="dnorm(x)", tf=TR, a=-1, b=1)
cont.trans(fx="dnorm(x)", tf=TR, a=-1, b=1, prt=FALSE, ws=c(9,3))
cont.trans("(x+1)/2", c(-1,1), TR, a=-0.3, b=0.7, ws=c(9,3))
# Non-linear Transformation
cont.trans(fx="dnorm(x)", tf=c(y="x^2",w="exp(x)"), a=-1, b=1, ws=c(9,3))
TR <- c(y="log(x+2)", w="x^2")
cont.trans("(x+1)/2", c(-1,1), TR, a=-0.3, b=0.7, ws=c(9,3))

}
