% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{corr.tsp}
\alias{corr.tsp}
\title{Pearson & Spearman Correlation Coefficient}
\usage{
corr.tsp(
  x,
  y,
  r0 = 0,
  alp = 0.05,
  alt = "two",
  type = "s",
  dig = 4,
  ws = c(8, 4)
)
}
\arguments{
\item{x}{Data of the first variable.}

\item{y}{Data of the second variable.}

\item{r0}{Correlation coefficient value under H0, Default: 0.}

\item{alp}{Level of significance, Default: 0.05.}

\item{alt}{Type alternative hypothesis, Default: 'two'.}

\item{type}{Calculation method: one of ("s", "p", "b"), Default: "s".}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(8,4).}
}
\value{
Test statistics.
}
\description{
Pearson Correlation Coefficient & Spearman Correlation Coefficient.
}
\examples{
x <- c(10,7,0,1,5, 2,8,6,4,9, 3,0,2,4,6, 8)
y <- c(75,77,91,64,79, 81,90,86,82,76, 89,93,80,87,83, 78)
corr.tsp(x, y)
corr.tsp(x, y, alt="le")
}
