% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{cprt}
\alias{cprt}
\title{Conditional Probability}
\usage{
cprt(A, B, prt = TRUE, dig = 0)
}
\arguments{
\item{A}{Event to be conditioned.}

\item{B}{Event to be conditioning.}

\item{prt}{Logical value for printing detailed output, Default: TRUE.}

\item{dig}{The number of decimal places, Default: 0 (fraction).}

\item{An}{The name of the event to be conditioned (cprt2).}

\item{Bn}{The name of the event to be conditioning (cprt2).}
}
\value{
The conditional probability.
}
\description{
Calculating the Conditional Probability
}
\examples{
S <- rolldie2(2)
A <- subset(S, (X1+X2) >=8)
B <- subset(S, (X1==6 | X2==6))
cprt(A, B)
cprt2(A, An="Sum >=8", B, Bn="At least one 6")
}
