% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{cprt3}
\alias{cprt3}
\title{Conditional Probability for the Events from Sequential Experments}
\usage{
cprt3(A, N1, B, N2, prt = TRUE, dig = 0)
}
\arguments{
\item{N1}{The size of the sample space of the entire experiment.}

\item{N2}{The size of the sample space of the initial experiment.}

\item{prt}{Logical value for printing detailed output, Default: TRUE.}

\item{dig}{The number of decimal places, Default: 0 (fraction).}

\item{a}{Event in the entire experiment.}

\item{b}{Event in the initial experiment.}
}
\value{
The conditional probability.
}
\description{
Calculating the Conditional Probability
}
\examples{
S <- rolldie2(2)
A <- subset(S, (X1==6 | X2==6))
B <- subset(S, (X1+X2) >=8)
cprt(B, A)
}
