% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{cross.test}
\alias{cross.test}
\title{Chi-square Test for Cross Tabulation Analysis}
\usage{
cross.test(x, v1, v2, alp = 0.05, dtab = 2, dig = 4, ws = c(7, 4))
}
\arguments{
\item{x}{Cross Table of frequency.}

\item{v1}{Row variable name.}

\item{v2}{Column variable name.}

\item{alp}{Level of significance, Default: 0.05.}

\item{dtab}{Number of decimal places for the table, Default: 2.}

\item{dig}{Number of decimal places for other output, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}
}
\value{
list(stat, df, cv, pv, exp, ress)
}
\description{
Chi-square test of homogeneity or independence for a cross table.
}
\examples{
x <- c(39,18,12,31,14, 35,23,18,35,13,  27,16,17,24,8,  9,12,8,19,22)
x <- matrix(x, nrow=4, ncol=5, byrow=TRUE)
rownames(x) <- c("Kor", "Eng", "Math", "Etc")
colnames(x) <- c("Sam", "Pub", "Exp", "Sal", "Etc")
cross.test(x)

data(Hospital, package="vcd")
cross.test(Hospital)

}
