% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{cwruns}
\alias{cwruns}
\title{Critical Values of the Two Sample Runs Test}
\usage{
cwruns(n1, n2, alpha = 0.05)
}
\arguments{
\item{n1}{Number of data in group 1 (scalar).}

\item{n2}{Number of data in group 2 (scalar).}

\item{alpha}{The level of significance, Default: 0.05.}
}
\value{
The lower and upper sided critical values.
}
\description{
The Lower and Upper Sided Critical Values of the Two Sample Runs Test.
}
\examples{
cwruns(5, 5)
cwruns(5, 5, alpha=0.025)

cwruns(5, 10)
cwruns(5, 10, alpha=0.025)
}
