% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{disc.cdf}
\alias{disc.cdf}
\title{CDF of a Discrete Random Variable}
\usage{
disc.cdf(fx, prt = TRUE, dig = 4, ws = c(7, 5), ...)
}
\arguments{
\item{fx}{Frequency table of the discrete random variable.}

\item{prt}{Logical: print the distribution? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,5).}

\item{...}{Other graphic parameters.}
}
\value{
list(CDF=xcdf, mean=EX, sd=DX).
}
\description{
Cumulative Distribution Function of a Discrete Random Variable.
}
\examples{
# CDF of the sum of heads in five coin tosses
coin5 <- choose(5, 0:5)
names(coin5) <- 0:5
disc.cdf(coin5, main="CDF of the Sum of Five Coins Heads")

# CDF of the random variable of B(10, 0.7)
B10.7 <- dbinom(0:10, 10, 0.7)
names(B10.7) <- 0:10
disc.cdf(B10.7, cex=0.8)
}
