% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch6-fn.R
\name{disc.dist}
\alias{disc.dist}
\title{Expected Values and Probabilities of Discrete Distributions}
\usage{
disc.dist(dist, par1, par2, nd, lb, ub, prt = "exp", dig = 4, ws = "n", ...)
}
\arguments{
\item{dist}{Discrete probability distribution.}

\item{par1}{First parameter of the distribution.}

\item{par2}{Second parameter of the distribution (if necessary).}

\item{nd}{Sample size for the hypergeometric distribution}

\item{lb}{Lower bound for findind P(lb <= X).}

\item{ub}{Upper bound for findind P(X <= ub).}

\item{prt}{Print option: one of c("exp", "var", "n"), Default: "exp".}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}

\item{...}{Other graphic parameters.}
}
\value{
list(EX=expected value, VX=variance, Pr=probability)
}
\description{
Expected Values and Probabilities of Discrete Distributions
}
\examples{
# Discrete uniform
disc.dist("unif", 20, lb=15, prt="var")

# Binomial
disc.dist("binom", 30, 0.05, prt="var")
disc.dist("binom", 30, 0.05, prt="p", lb=3, ws=c(7,4), xlim=c(0,10))

}
