% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{disc.jexp}
\alias{disc.jexp}
\title{Expected Values of Two Discrete Random Variables}
\usage{
disc.jexp(X, Y, FUN, Tr, prt = "exp", dig = 4, detail = FALSE)
}
\arguments{
\item{X}{Joint frequency (probability) table or vector.}

\item{Y}{The second random variable (required if X is a vector).}

\item{FUN}{Joint frequency (probability) function (if necessary).}

\item{Tr}{Transformation formula in string.}

\item{prt}{Print option: one of c("exp", "var", "cov", "cor"), Default: "exp".}

\item{dig}{Number of decimal places, Default: 4.}

\item{detail}{Logical: print the joint & marginal PDF? Default: FALSE.}
}
\value{
list(EX=E(X), EY=E(Y), VX=V(X), VY=V(Y), EXY=E(XY), VXY=Cov(X,Y), CXY=Corr(X,Y))
}
\description{
Joint Probabilities and Expected Values of Two Discrete Random Variables
}
\examples{

# [1] Input the elements of X and Y, respectively
S <- rolldie2(4)
X <- apply(S, 1, max)
Y <- apply(S, 1, min)
disc.jexp(X, Y, prt="cov")
disc.jexp(X, Y, prt="cor", detail=TRUE)

S = rolldie2(3)
X = apply(S, 1, median)
Y = apply(S, 1, \(x) max(x)-min(x))
disc.jexp(X, Y, prt="cor")

# [2] Functional input
fxy = function(x, y) x+y
x=0:4; y=0:3
disc.jexp(y, x, fxy, prt="cor")

# [3] Tabular input
tab = with(mtcars, table(cyl, gear))
disc.jexp(tab, prt="cor")

}
