% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{disc.jpdf}
\alias{disc.jpdf}
\title{Joint PDF of Two Discrete Random variable}
\usage{
disc.jpdf(
  X,
  Y,
  FUN,
  type = "j",
  xc,
  yc,
  prt = TRUE,
  dig = 4,
  ws = "n",
  cex = 0.8
)
}
\arguments{
\item{X}{Sample space (or range) of X, or table(X,Y).}

\item{Y}{Sample space (or range) of Y (optional).}

\item{FUN}{Joint PDF of X and Y (optional).}

\item{type}{"j"=joint, "m"=margin, "c"=condition, "i"=independ, Default: "j".}

\item{xc}{Given conditioning value of X (type="c").}

\item{yc}{Given conditioning value of Y (type="c").}

\item{prt}{Logical: print frequency table? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}

\item{cex}{Text size of f(x,y) in the plot, Default: 0.8.}
}
\value{
Joint frequency and probability table.
}
\description{
Joint, Marginal, Conditional Probability Distribution of Two Discrete Random variables.
}
\examples{
require(scatterplot3d)
# [1] Input the elements of X and Y, respectively
S = rolldie2(3)
X = apply(S, 1, median)
Y = apply(S, 1, \(x) max(x)-min(x))
# Joint
disc.jpdf(X, Y)
disc.jpdf(X, Y, prt=FALSE, ws=c(7,5))
# Marginal
disc.jpdf(X, Y, type="m")
disc.jpdf(X, Y, type="m", prt=FALSE, ws=c(7,5))
# Conditional
disc.jpdf(X, Y, type="c", yc=0:5)
disc.jpdf(X, Y, type="c", yc=0:5, prt=FALSE, ws=c(7,5))
# Independence
disc.jpdf(X, Y, type="i")

# [2] Functional input
fxy = function(x, y) x^2+y
x=1:3; y=1:5
disc.jpdf(x, y, fxy, ws=c(7,5))

# Marginal
disc.jpdf(x, y, fxy, type="m", ws=c(7,5))
# Conditional
disc.jpdf(x, y, fxy, type="c", xc=1:3, ws=c(9,3))
# Independence
disc.jpdf(x, y, fxy, type="i")

# [3] Tabular input
x=1:3; y=1:5
tab = outer(x, y, fxy)
dimnames(tab) <- list(X=x, Y=y)
disc.jpdf(tab, ws=c(7,5))

# Marginal
disc.jpdf(tab, type="m", ws=c(7,5))
# Conditional
disc.jpdf(tab, type="c", xc=1:3, ws=c(9,3))
# Independence
disc.jpdf(tab, type="i")

}
