% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch6-fn.R
\name{disc.mexp}
\alias{disc.mexp}
\title{Expected Values and Plots of Multiple Discrete PDFs}
\usage{
disc.mexp(xv, fx, fx2, fx3, prt = "var", dig = 4, ws = "n", del = 0.2, ...)
}
\arguments{
\item{xv}{Values of the random variable.}

\item{fx}{List of the first PDFs (2~9).}

\item{fx2}{List of the second PDFs (same number of fx).}

\item{fx3}{List of the third PDFs (same number of fx).}

\item{prt}{Print option: one of c("var", "exp", "n"), Default: "var".}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}

\item{del}{Distance for overlapping plots, Default: 0.2.}

\item{...}{Other graphic parameters.}
}
\value{
list(Ex, Vx, Dx)
}
\description{
Expected Values and Plots of Multiple Discrete PDFs.
}
\examples{
# Binomial distributions
n <- 10;  p <- 1:9/10; x <- 0:n
fx1 <- fx2 <- list()
for (i in 1:9) fx1[[i]] <- dbinom(x, n, p[i])
mt1 <- paste0("B(10,", p, ")")
disc.mexp(x, fx1, prt="n", ws=c(8,6), main=mt1, lwd=5)

# Binomial vs. Hypergeometric
N <- 50;  S <- 1:9*5; n <- 10; x <- 0:n
for (i in 1:9) fx2[[i]] <- dhyper(x, S[i], N-S[i], n)
mt12 <- paste0("HG(10,50,", S,"):B(10,", p,")")
disc.mexp(x, fx2, fx1, ws=c(8,6), main=mt12)
}
