% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{dwruns}
\alias{dwruns}
\title{PMF of the Number of Runs}
\usage{
dwruns(r, n1, n2)
}
\arguments{
\item{r}{Number of Runs (vector).}

\item{n1}{Number of data in group 1 (scalar).}

\item{n2}{Number of data in group 2 (scalar).}
}
\value{
P(R=r).
}
\description{
The Probabilityt Mass Function of the Number of Runs.
}
\examples{
dwruns(2:10, 5, 5)
sum(dwruns(2:10, 5, 5))
dwruns(1:12, 5, 5)
sum(dwruns(1:12, 5, 5))

dwruns(2:11, 5, 10)
sum(dwruns(2:11, 5, 10))
dwruns(1:13, 5, 10)
sum(dwruns(1:13, 5, 10))
}
