% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{element}
\alias{element}
\title{Elements of an Event}
\usage{
element(A, ncol = 10, prt = TRUE)
}
\arguments{
\item{A}{Event in data frame.}

\item{ncol}{Maximum number of elements in each line, Default: 10.}

\item{prt}{Logical value for printing elements, Default: TRUE.}
}
\value{
Vector of elements of the event.
}
\description{
Displaying Elements of an Event
}
\examples{
S <- rolldie2(2)
B <- subset(S, (X1+X2) >=8)
element(B)

S <- tosscoin2(4)
S = S[order(S[[1]], S[[2]], S[[3]], S[[4]]), ]
A <- subset(S, apply(S, 1, function(x) sum(x=="H") >=2))
element(A, ncol=6)
}
