% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{f.prob}
\alias{f.prob}
\title{Cumulative Probability of the F-distribution}
\usage{
f.prob(nu1, nu2, ub, lb = 0, prt = TRUE, dig = 4, ws = c(7, 4), ...)
}
\arguments{
\item{nu1}{Degree of freedom of the numerator.}

\item{nu2}{Degree of freedom of the denominator.}

\item{ub}{Values of upper limits for F(ub).}

\item{lb}{Values for F(ub)-F(lb), Default: 0.}

\item{prt}{Logical: print the output? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Grapic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
Cumulative probabilities.
}
\description{
Cumulative Probability of the F-distribution
}
\examples{
k <- 1:7
nu1 <- 8; nu2 <- 5
f.prob(nu1, nu2, ub=k)

f.prob(nu1, nu2, lb=0:3, ub=7:10)

}
