% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{f.quant}
\alias{f.quant}
\title{Quantile Plot of the F-distribution}
\usage{
f.quant(nu1, nu2, pv, prt = TRUE, dig = 4, ws = c(7, 4), PDF = TRUE, ...)
}
\arguments{
\item{nu1}{Degree of freedom of the numerator.}

\item{nu2}{Degree of freedom of the denominator.}

\item{pv}{Probability values to find the quantiles.}

\item{prt}{Logical: print the output? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Grapic window size, Default: c(7,4).}

\item{PDF}{Logical: plot the PDF? (o.w. CDF) Default: TRUE.}

\item{...}{Other graphic parameters.}
}
\value{
Quantiles.
}
\description{
Quantile Plot of the F-distribution.
}
\examples{
nu1 <- 8; nu2 <- 5
pv <- c(0.005, 0.01, 0.025, 0.05, 0.5, 0.95, 0.975, 0.99, 0.995)
f.quant(nu1, nu2, pv, pv2)
}
