% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{fdist.sim}
\alias{fdist.sim}
\title{Simulation of the F-distribution}
\usage{
fdist.sim(
  sig1 = 1,
  sig2 = 1,
  n1,
  n2,
  N = 1000,
  prt = TRUE,
  ws = c(10, 4),
  dig = 4,
  ...
)
}
\arguments{
\item{N}{Number of random values, Default: 10000}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{nu1}{Numerator degree of freedom, Default: 5}

\item{nu2}{Denominator degree of freedom, Default: 5}

\item{ng}{Number of classes in histogram, Default: 250}

\item{seed}{Seed value for random number generator, Default: 9857}

\item{xp}{Vector of x-axis values, Default: 1:9}
}
\value{
None.
}
\description{
Simulation of the F-distribution
}
\examples{
fdist.sim(nu1=8, nu2=5)

fdist.sim(50, 50, ng=100, xp=1:5/2)
}
