% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{friedman.tsp}
\alias{friedman.tsp}
\title{Friedman Test}
\usage{
friedman.tsp(x, a, b, dig = 4, ws = "n")
}
\arguments{
\item{x}{Response data.}

\item{a}{Data of factor levels.}

\item{b}{Data of block levels.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}
}
\value{
None.
}
\description{
Friedman Test with a Plot
}
\examples{
x <- c(71,77,75,79,88, 70,94,74,74,83, 72,95,77,80,90, 94,64,76,76,89)
a <- rep(1:4, each=5)
b <- rep(1:5, 4)
friedman.tsp(x, a, b, ws=c(7,4))
}
