% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{ftest.plot}
\alias{ftest.plot}
\title{Plot the PDF of the F-test Statistic}
\usage{
ftest.plot(F0, df, nalt, dig = 4, leg, ...)
}
\arguments{
\item{fstat}{F-test statistic for the ratio of two population variances}

\item{deg}{Vector of degree of freedoms}

\item{alt}{Type of the alternative hypothesis, Default: 'two'}

\item{pmax}{Coverage probability of x-axis, Default: 0.995}

\item{prt}{Logical value for printing the p-value, Default: FALSE}
}
\value{
None.
}
\description{
Plot the PDF of the F-test Statistic
}
\examples{
set.seed(1234)
x <- rnorm(20, 10, 2.4); y <- rnorm(25, 12, 1.8)
(vo <- var.test(x, y))
win.graph(7, 5)
ftest.plot(stat=vo$stat, deg=as.vector(vo$para), alt="two", pmax=0.999)
}
