% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{inter.count}
\alias{inter.count}
\title{Count the Elements of Intersections}
\usage{
inter.count(S, EV, subid)
}
\arguments{
\item{S}{Sample Space (data frame or vector).}

\item{EV}{List of events (data frame or vector).}

\item{subid}{Code for specific events to apply.}

\item{name}{Logical value for using the event names, Default:FALSE.}
}
\value{
The number of elements for all partitions in a data frame.
}
\description{
Make a Data frame of Counting the Elements of Intersections.
}
\examples{
S <- rolldie2(4); N <- nrow(S)
A <- subset(S, ((X1+X2)\%\%2) == 0)
B <- subset(S, (X1+X2) >= 8)
C <- subset(S, abs(X1-X2) <= 1)
D <- subset(S, apply(S, 1, max) == 6)
E <- subset(S, apply(S, 1, min) == 1)
inter.count(S, list(A,B,C,D,E))
# Count ABD^E
inter.count(S, EV=list(A,B,D,E), subid=c(1,1,0,1))
}
