% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{lln.coin}
\alias{lln.coin}
\title{The Law of Large Numbers: Tossing a Coin}
\usage{
lln.coin(pr = 0.5, np = 30, nmax = 100, int = 0.01, ws = c(7, 4), col = "blue")
}
\arguments{
\item{pr}{The probability of the head, Default: 0.5.}

\item{np}{Times for sampling in each run, Default: 30.}

\item{nmax}{Maximum number of runs.}

\item{int}{Time interval of runs per second, Default: 0.01.}

\item{ws}{Size of graphic window, Default: c(7,4).}

\item{col}{Color of the polygons of estimates, Default: "blue".}
}
\value{
The number of elements for all partitions in a data frame.
}
\description{
Animate the Law of Large Numbers for Tossing a Coin.
}
\examples{
lln.coin(nmax=500)
lln.coin(np=5, nmax=500)
lln.coin(pr=0.8, col="magenta")
lln.coin(pr=0.8, nmax=1000)
}
