% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{mean1.ci}
\alias{mean1.ci}
\title{Confidence Interval for a Mean}
\usage{
mean1.ci(x, sig, n, pvar = FALSE, alp = 0.05, dig = 4)
}
\arguments{
\item{x}{Sample mean, or sample data}

\item{sig}{Standard deviation (unnecessary when sample data are given)}

\item{n}{Sample size (unnecessary when sample data are given)}

\item{pvar}{Logical value for known population variance, Default: FALSE}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Confidence Interval for a Single Population Mean
}
\examples{
mean1.ci(x=199.5, sig=5, n=20, pvar=TRUE)
mean1.ci(x=199.5, sig=5, n=20)

set.seed(1234)
x <- rnorm(20, 199.5, 5)
mean1.ci(x, sig=5, pvar=TRUE)
mean1.ci(x)

}
