% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{mean2.test}
\alias{mean2.test}
\title{Inference on the Difference of Two Population Means}
\usage{
mean2.test(
  x,
  y,
  alt = "two",
  alp = 0.05,
  sig,
  n,
  d0 = 0,
  var = "eq",
  dig = 4,
  ws = c(7, 4),
  ...
)
}
\arguments{
\item{x}{First sample data or sample mean.}

\item{y}{Second sample data or sample mean.}

\item{alt}{Type of the alternative hypothesis, Default: 'two'.}

\item{alp}{Level of significance, Default: 0.05.}

\item{sig}{Standard deviations of two groups (without sample data).}

\item{n}{Sample size of two groups (without sample data).}

\item{d0}{Difference of two population means under H0, Default: 0.}

\item{var}{Status of variance (one of "eq", "uneq", "known"), Default: 'eq'.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
Summary statistics.
}
\description{
Hypothesis Test and CI for the Difference of Two Population Means
}
\examples{
mean2.test(x=198.5, y=201.3, sig=5, n=c(25,34), var="known")
mean2.test(x=198.5, y=201.3, sig=5, n=30, var="known")
mean2.test(x=198.5, y=201.3, sig=c(4.8,5.1), n=c(25,34), var="eq")
mean2.test(x=198.5, y=201.3, sig=c(2.8,5.5), n=c(25,34), var="uneq")

set.seed(3691)
x <- rnorm(20, 199, 2)
y <- rnorm(25, 200, 4)
mean2.test(x, y, var="eq")
mean2.test(x, y, var="uneq")
}
