% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{mosaic.res}
\alias{mosaic.res}
\title{Mosaic Plot with Pearson Residuals for a Cross Table}
\usage{
mosaic.res(tab, main, dig = 4, resid = TRUE, ws = c(7, 6))
}
\arguments{
\item{tab}{Cross table of frequency.}

\item{main}{Graph title.}

\item{dig}{Number of decimal places, Default: 4.}

\item{resid}{Logical: display Pearson residuals? Default: TRUE.}

\item{ws}{Graphic window size, Default: c(7,6).}
}
\value{
Object from chisq.test()
}
\description{
Mosaic plot for a cross table displaying Pearson residuals.
}
\examples{
x <- c(39,18,12,31,14, 35,23,18,35,13,  27,16,17,24,8,  9,12,8,19,22)
x <- matrix(x, nrow=4, ncol=5, byrow=TRUE)
Subject <- c("Kor", "Eng", "Math", "Etc")
Hope <- c("Sam", "Pub", "Exp", "Sal", "Etc")
t1 <- as.table(x)
dimnames(t1) <- list(Subject=Subject, Hope=Hope)
# require(vcd)
mosaic.res(tab=t1, main="Students' Favorite Subject and Hope")

data(Hospital, package="vcd")
mosaic.res(Hospital)
}
