% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{mreg.diff}
\alias{mreg.diff}
\title{Compare Two Multiple Regression Models}
\usage{
mreg.diff(form1, form2, data, alp = 0.05, dig = 4, dp = c(1:3, 5))
}
\arguments{
\item{form1}{Formula of a regression model (ex: y ~ x1 + x2)}

\item{form2}{Formula for another regression model (eg: y ~ x1 * x2)}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{plot}{Logical value for drawing diagnosis plots, Default: TRUE.}

\item{wh}{Index vector for the diagnosis plots, Default: c(1:3,5).}
}
\value{
ANOVA table.
}
\description{
Compare Two Multiple Regression Models.
}
\examples{
attach(mtcars)
form.1 <- mpg ~ hp + drat + wt
form.2 <- mpg ~ hp * drat * wt
mreg.diff(form.1, form.2)

data(exa14_10)
Score <- exa14_10[[1]]; Study <- exa14_10[[2]]; Reading <- exa14_10[[3]]
form2.1 <- Score ~ Study + Reading
form2.2 <- Score ~ Study * Reading
mreg.diff(form2.1, form2.2)
}
