% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{mreg.pred}
\alias{mreg.pred}
\title{Prediction and Confidence Interval: Multiple Regression}
\usage{
mreg.pred(form, data, newd, alp = 0.05, dig = 4, ws = "n", pvx = 1, ...)
}
\arguments{
\item{form}{Formula of regression model (ex: y ~ x1 + x2)}

\item{newd}{A new data frame of independent variables.}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{pvx}{Order of selected independent variables, Default: 1.}

\item{plot}{Logical value for plotting CI and PI, Default: FALSE.}

\item{xrng}{Plot range of the independent variable.}

\item{nx}{Number of plotting points, Default: 50.}
}
\value{
CI and PI
}
\description{
Prediction and Confidence Interval in a Multiple Regression Model.
}
\examples{
# Data set 'mtcars'
form <- mpg ~ hp + drat + wt
newd <- data.frame(hp=300, drat=4, wt=4)
mreg.pred(form, data=mtcars, newd)
mreg.pred(form, data=mtcars, ws=c(7,6), pvx=1, xlim=c(30,400))
mreg.pred(form, data=mtcars, ws=c(7,6), pvx=2, xlim=c(2,6))
mreg.pred(form, data=mtcars, ws=c(7,6), pvx=3, xlim=c(1,6))
# Example 14.9
data(exa14_9)
form2 <- Score ~ Study + Reading
nd <- data.frame(Study=5, Reading=5)
mreg.pred(form2, data=exa14_9, newd=nd)
mreg.pred(form2, data=exa14_9, ws=c(7,6), xrng=c(0,15), pvx=1)
mreg.pred(form2, data=exa14_9, ws=c(7,6), xrng=c(0,15), pvx=2)

}
