% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{mult.corr}
\alias{mult.corr}
\title{Correlation Coefficients and Scatter Plots}
\usage{
mult.corr(data, item, prt = TRUE, dig = 4, detail = FALSE, ws = "n", ...)
}
\arguments{
\item{data}{Input data frame (or list or matrix) of variables.}

\item{item}{Selected names of the variables.}

\item{prt}{Logical: print the result? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{detail}{Logical: print frequency tables? Default: FALSE.}

\item{ws}{Graphic window size, Default: "n".}

\item{...}{Othe graphic parameters.}
}
\value{
Corr(X,Y) table.
}
\description{
Correlation Coefficients and Scatter Plots of Random variables
}
\examples{
S <- prob::rolldie(3)
X = data.frame(Sum=apply(S, 1, sum), 
               Range=apply(S, 1, \(x) max(x)-min(x)),
               Max=apply(S, 1, max), 
               Min=apply(S, 1, min))
mult.corr(X, ws=c(7,6))

select = c("mpg", "disp", "hp", "drat", "wt", "qsec")
mult.corr(mtcars, select, ws=c(7,6))
}
