% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{mult.hist}
\alias{mult.hist}
\title{Multiple Histograms}
\usage{
mult.hist(data, breaks, ws, mfd, ...)
}
\arguments{
\item{data}{Data frame of input data.}

\item{breaks}{Break method Default: Sturges.}

\item{ws}{Graphic window size vector.}

\item{mfd}{Multi-frame dimension vector.}

\item{...}{Other graphic parameters.}

\item{mt}{Vector of main titles.}

\item{xl}{Vector of x labels.}

\item{col}{Vector of colors for the histograms.}

\item{mar}{Margin vector of the plot.}
}
\value{
None.
}
\description{
To draw multiple histograms from a data frame.
}
\examples{
mult.hist(iris[1:4])
mult.hist(iris[1:4], col=rainbow(4))
mult.hist(iris[1:4], col=rainbow(4), ws=c(12,4), mfd=c(1,4))
mult.hist(mtcars[1:7], col=rainbow(7))
}
\keyword{Ch2.}
\keyword{Descriptive}
\keyword{Statistics}
