% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch6-fn.R
\name{multinom.plot}
\alias{multinom.plot}
\title{Multinomial PDF Plots, Expectations, and Probabilities}
\usage{
multinom.plot(ps, size, lb, ub, prt = "exp", dig = 4, ws = "n", ...)
}
\arguments{
\item{ps}{Probability (or proportion) matrix with rows for multiple cases.}

\item{size}{Sample size.}

\item{lb}{Lower bound vector for finding a probability.}

\item{ub}{Upper bound vector for finding a probability.}

\item{prt}{Print option: one of c("var", "exp", "n"), Default: "exp".}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}

\item{...}{Other graphic parameters.}
}
\value{
Matrix of multinomial PDF
}
\description{
PDF Plots, Expectations, and Probabilities of Multinomial Distribution.
}
\examples{
# Comparing four multinomial distributions
ps <- matrix(c(1,1,8, 1,5,4, 4,4,2, 1,1,1), nrow=4, ncol=3, byrow=T)
multinom.plot(ps, 5)
multinom.plot(ps, 5, prt="var")
multinom.plot(ps, 5, prt="n", ws=c(7,6))
}
