% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{norm.sim}
\alias{norm.sim}
\title{Simulation of Normal Sample Means}
\usage{
norm.sim(
  ns,
  mu = 0,
  sig = 1,
  N = 1000,
  prt = TRUE,
  ws = c(10, 4),
  dig = 4,
  ...
)
}
\arguments{
\item{ns}{Sample size.}

\item{mu}{Expected value, Default: 0.}

\item{sig}{Standard deviation, Default: 1.}

\item{N}{Number of iterations, Default: 1000.}

\item{prt}{Logical: print statistics? Default: TRUE.}

\item{ws}{Graphic window size, Default: c(10,4).}

\item{dig}{Number of decimal places, Default: 4.}

\item{...}{Other graphic parameters.}
}
\value{
Sample means and standard deviations.
}
\description{
Simulation for the Distribution of Normal Sample Means.
}
\examples{
norm.sim(ns=4, mu=5, sig=9, lwd=3)

norm.sim(ns=25, mu=100, sig=10, N=10000, pos="topright")

norm.sim(ns=10, mu=100, sig=10, ng=100, N=100000)
}
