% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{norm.spn}
\alias{norm.spn}
\title{Minimum Number of Samples to Meet an Error Limit}
\usage{
norm.spn(kp, alp, prt = TRUE, ws = c(7, 4), log = TRUE, ...)
}
\arguments{
\item{kp}{Error limits in multiples of the standard deviation.}

\item{alp}{Levels of significance.}

\item{prt}{Logical: print the sample size? Default: TRUE.}

\item{ws}{Graphic window size, Default: c(7,4).}

\item{log}{Logical: use log-scale for y-axis? Default: TRUE.}

\item{...}{Other graphic parameters.}
}
\value{
Sample sizes.
}
\description{
Minimum Number of Samples to Meet an Error Limit of Sample Means from Normal Population.
}
\examples{
alp <- c(0.01, 0.05, 0.1)
dcol <- c("red", "blue", "green4")
norm.spn(kp=0.4, alp, col=dcol)

norm.spn(kp=0.05, alp, cex=0.8, col=dcol)

norm.spn(kp=1:5/10, alp, col=dcol)

norm.spn(kp=0.2, alp=1:10/100,cex=0.8,col="cyan",xlim=c(0.15,0.25))

}
