% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{normtest.plot}
\alias{normtest.plot}
\title{Plot the PDF of Test Statistic under the Normal Distribution}
\usage{
normtest.plot(z0, alt = "two", dig = 4, mt, prng = c(-4, 4), xlab, prt = FALSE)
}
\arguments{
\item{z0}{Z-test statistic for the population mean}

\item{alt}{Type of the alternative hypothesis, Default: 'two'}

\item{dig}{Number of decimal places, Default: 4.}

\item{mt}{Main title of the plot}

\item{prng}{Range of x-axis, Default: c(-4, 4)}

\item{xlab}{Label of x-axis, Default: 'Test Statistic (Z0)'}

\item{prt}{Logical value for printing the p-value, Default: FALSE}
}
\value{
None.
}
\description{
Plot the PDF of Test Statistic under the Normal Distribution
}
\examples{
win.graph(7, 5)
normtest.plot(1.93)
normtest.plot(-1.93, alt="le")
}
