% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{power.plot}
\alias{power.plot}
\title{Power Function of the Test for a Mean}
\usage{
power.plot(
  mu0,
  mu1,
  sig,
  nv,
  alt,
  alp = 0.05,
  pwr = TRUE,
  dig = 4,
  ws = c(7, 4),
  ...
)
}
\arguments{
\item{mu0}{Population mean under H0, Default: 0.}

\item{mu1}{Population mean under H1.}

\item{sig}{Population standard deviation, Default: 1.}

\item{nv}{Sample size vector, Default: c(10,30,50,100).}

\item{alt}{Type of the alternative hypothesis ("gr", "le", "two").}

\item{alp}{Level of significance, Default: 0.05.}

\item{pwr}{Logical: apply power function? (o.w. OC curve) Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
Power function (or OC function).
}
\description{
Power Function of the Test for a Population Mean.
}
\examples{
power.plot()
power.plot(pwr=FALSE)
power.plot(mu1=0.3)
power.plot(mu1=0.3, pwr=FALSE)

power.plot(mu0=100, mu1=102, sig=5, alt="gr")
power.plot(mu0=100, mu1=102, sig=5, alt="gr", pwr=FALSE)
n = c(10, 30, 50, 70, 100)
power.plot(mu0=100, mu1=102, sig=5, nv=n)
power.plot(mu0=100, mu1=102, sig=5, nv=n, pwr=FALSE)

power.plot(nv=50:150)
power.plot(nv=50:150, pwr=FALSE)

}
