% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{pprt}
\alias{pprt}
\title{Probability of an Event}
\usage{
pprt(A, N, prt = TRUE, dig = 0)
}
\arguments{
\item{A}{Event in data frame.}

\item{N}{Size of the sample space.}

\item{prt}{Logical value for printing detailed output, Default: TRUE.}

\item{dig}{The number of decimal places, Default: 0 (fraction).}

\item{An}{The name of the event (pprt2).}
}
\value{
The probability of an event.
}
\description{
Calculating the Probability of an Event
}
\examples{
S <- rolldie2(2)
B <- subset(S, (X1+X2) >=8)
pprt(B, nrow(S))
pprt2(B, An="Sum >= 8", nrow(S))

S <- tosscoin2(4)
A <- subset(S, apply(S, 1, function(x) sum(x=="H") >=2))
pprt(A, nrow(S))
pprt2(A, "Heads >= 2", nrow(S))
}
