% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{prop1.ci}
\alias{prop1.ci}
\title{Confidence Interval for a Population Proportion}
\usage{
prop1.ci(x, n, alp = 0.05, prt = TRUE, dig = 4, method = "all")
}
\arguments{
\item{x}{Number of successes in a sample}

\item{n}{Sample size}

\item{alp}{Level of significance, Default: 0.05}

\item{prt}{Logical value for printing the output, Default: TRUE}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{method}{One of ("all", "normal", "wilson", "binom"), Default: "all"}
}
\value{
c(lcl, ucl)
}
\description{
Confidence Interval for a Population Proportion by Three Methods
}
\examples{
prop1.ci(n=200, x=15)
prop1.ci(n=200, x=11:15)
prop1.ci(n=200, x=15, alp=c(0.1, 0.05, 0.01))
}
