% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{prop2.test}
\alias{prop2.test}
\title{Approximate Test for the Difference of Two Proportions (Large Sample)}
\usage{
prop2.test(
  x,
  y,
  n1,
  n2,
  alt = "two",
  alp = 0.05,
  d0 = 0,
  dig = 4,
  ws = c(7, 4),
  ...
)
}
\arguments{
\item{x}{Number of successes in the first sample data.}

\item{y}{Number of successes in the second sample data.}

\item{n1}{First sample size.}

\item{n2}{Second sample size.}

\item{alt}{Type of the alternative hypothesis, Default: "two".}

\item{alp}{Level of significance, Default: 0.05.}

\item{d0}{Difference of two proportions under H0, Default: 0.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
Summary statistics.
}
\description{
Approximate Test for the Difference of Two Population Proportions (Large Sample).
}
\examples{
prop2.test(x=12, y=10, n1=150, n2=250)
prop2.test(x=12, y=10, n1=150, n2=250, alt="gr")
}
