% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{pwruns}
\alias{pwruns}
\title{CDF of the Number of Runs}
\usage{
pwruns(r, n1, n2, lower = TRUE)
}
\arguments{
\item{r}{Number of Runs (vector).}

\item{n1}{Number of data in group 1 (scalar).}

\item{n2}{Number of data in group 2 (scalar).}

\item{lower}{Logical value for the lower tail, Default: TRUE.}
}
\value{
P(R<=r) or P(R>r) when lower=FALSE.
}
\description{
The Cumulative Distribution Function of the Number of Runs.
}
\examples{
pwruns(1:12, 5, 5)
pwruns(1:12, 5, 5, lower=FALSE)
pwruns(1:12, 5, 5) + pwruns(1:12, 5, 5, lower=FALSE)

pwruns(1:13, 5, 10)
pwruns(1:13, 5, 10, lower=FALSE)
pwruns(1:13, 5, 10) + pwruns(1:13, 5, 10, lower=FALSE)
}
