% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{ranksum.dist}
\alias{ranksum.dist}
\title{Distribution of Rank-Sum Test Statistic}
\usage{
ranksum.dist(n1 = 1:2, n2 = 2, alp = "n", dig = 4, ws = c(7, 4))
}
\arguments{
\item{n1}{Number of data in group 1, Default: 1:2.}

\item{n2}{Number of data in group 2, Default: 2.}

\item{alp}{Level of significance, Default: "n"}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}
}
\value{
None.
}
\description{
Distribution of Wilcoxon-Mann-Whitney Rank-Sum Test Statistic
}
\examples{
ranksum.dist(n1=1:3, n2=3:5, ws="n")
ranksum.dist(n1=c(2,4,6,8), n2=10, alp=0.05)
}
