% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{ranksum.tsp}
\alias{ranksum.tsp}
\title{Wilcoxon-Mann-Whitney Rank-Sum Test}
\usage{
ranksum.tsp(x, y, alt = "two", dig = 4, ws = c(7, 4))
}
\arguments{
\item{x}{Data of the first group.}

\item{y}{Data of the second group.}

\item{alt}{Type of alternative hypothesis, Default: 'two'.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}
}
\value{
None.
}
\description{
Wilcoxon-Mann-Whitney Rank-Sum Test with a Plot.
}
\examples{
x <- c(1,5,7,8,8,8,9)
y <- c(2,8,8,8,8,9,9,9,9,10,10,10,10)
ranksum.tsp(x, y)
ranksum.tsp(y, x)
}
