% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{rolldie.sum}
\alias{rolldie.sum}
\title{PDF of the Sum of n Dice}
\usage{
rolldie.sum(n, nf = 6, prt = FALSE, dig = 4, ws = c(7, 5), cex)
}
\arguments{
\item{n}{Number of die rolls.}

\item{nf}{Number of the die faces, Default: 6.}

\item{prt}{Logical: print the distribution? Default: FALSE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,5).}

\item{cex}{Text size in the plot.}
}
\value{
List of frequency and probability distribution of the sum.
}
\description{
Probability Distribution of the Sum of n Dice.
}
\examples{
# Rolling four dice
rolldie.sum(4)
# Rolling ten dice
dist <- rolldie.sum(10)
# The sample space has 60466176 elements.
sum(dist$freq); 6^10
}
