% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{runs.dist}
\alias{runs.dist}
\title{Plots and Tables of Runs Distribution}
\usage{
runs.dist(n1 = 2:5, n2 = 2:5, alp = 0.05, ws = c(7, 4))
}
\arguments{
\item{n1}{Number of data in group 1, Default: 2:5.}

\item{n2}{Number of data in group 2, Default: 2:5.}
}
\value{
None.
}
\description{
Plots and Tables of the Prbability Mass Function of Runs.
}
\examples{
runs.dist(n1=c(5,20), n2=c(5,20))
runs.dist(n1=5, n2=2:10)
runs.dist(n1=1:9*5, n2=9:1*5)
runs.dist(n1=1:10*5, n2=10:1*5)
}
