% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{runs.pval}
\alias{runs.pval}
\title{P-value of the Two Sample Runs Test}
\usage{
runs.pval(r, n1, n2, alt = "two")
}
\arguments{
\item{r}{Number of Runs (vector).}

\item{n1}{Number of data in group 1 (scalar).}

\item{n2}{Number of data in group 2 (scalar).}

\item{alt}{Direction of the alternative hypothesis, Default: 'two'.}
}
\value{
The p-value
}
\description{
The P-value of the Two Sample Runs Test
}
\examples{
runs.pval(1:12, 5, 5)
runs.pval(2:5, 5, 5) - runs.pval(10:7, 5, 5)
runs.pval(1:12, 5, 5, alt="le")
runs.pval(1:12, 5, 5, alt="gr")
runs.pval(2:10, 5, 5, alt="le") - runs.pval(10:2, 5, 5, alt="gr")

runs.pval(1:13, 5, 10)
runs.pval(1:13, 5, 10, alt="le")
runs.pval(1:13, 5, 10, alt="gr")
}
