% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{runs.table}
\alias{runs.table}
\title{Critical Values of Runs Test}
\usage{
runs.table(n1 = 2:5, n2 = 2:5, alp = 0.05)
}
\arguments{
\item{n1}{Number of data in group 1, Default: 2:5}

\item{n2}{Number of data in group 2, Default: 2:5}

\item{alp}{Level of significance, Default: 0.05}
}
\value{
None.
}
\description{
Lower and Upper Critical Values of Runs Test
}
\examples{
runs.table(n1=10, n2=10)
runs.table(n1=2:10, n2=2:10)
runs.table(n1=c(5,20), n2=c(5,20))
}
