% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{runs.tsp}
\alias{runs.tsp}
\title{Runs Test}
\usage{
runs.tsp(x, n1, n2, alp = 0.05, alt = "two", dig = 4, ws = c(7, 4))
}
\arguments{
\item{x}{Data vector (or number of runs).}

\item{n1}{Number of data in group 1 (without raw data).}

\item{n2}{Number of data in group 2 (without raw data).}

\item{alp}{Level of significance, Default: 0.05.}

\item{alt}{Type of alternative hypothesis, Default: 'two'.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}
}
\value{
None.
}
\description{
Runs Test with a Plot.
}
\examples{
x <- c(1,1,0,0,1,0,rep(1,7), rep(0,7))
runs.tsp(x)
runs.tsp(x, alt="le")

set.seed(1234); x <- rbinom(50, 1, 0.5)
runs.tsp(x)

x <- rep(0, 50)
x[c(1:2, 8:10, 17:18, 21, 26:27, 29:31, 36:37, 41:44, 49)] <- 1
runs.tsp(x)
}
