% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{scat.lm}
\alias{scat.lm}
\title{Scatter Plot with a Regression Line}
\usage{
scat.lm(x, y, data, dig = 2, ws, ...)
}
\arguments{
\item{x}{Vector of x variable.}

\item{y}{Vector of y variable.}

\item{dig}{The number of decimal places, Default: 4.}

\item{ws}{Graphic window size.}

\item{...}{Other graphic parameters.}
}
\value{
Object lm(y~x)
}
\description{
To create a scatter plot with a linear regression line.
}
\examples{
with(mtcars, scat.lm(wt, mpg))
scat.lm(wt, mpg, mtcars, main="Weight vs. MPG", dig=4, ws=c(7,5))

par(mfrow=c(2,2))
for (k in 3:6) scat.lm(mtcars[[k]], mtcars$mpg, 
                      xlab=names(mtcars)[k], ylab="mpg")
}
\keyword{Ch2.}
\keyword{Descriptive}
\keyword{Statistics}
