% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{signrank.dist}
\alias{signrank.dist}
\title{Plots and Tables of Wilcoxon Signed-Rank Distribution}
\usage{
signrank.dist(nv = 5:10, alp, ws = c(7, 4), sep = FALSE, pp)
}
\arguments{
\item{nv}{Number of paired samples, Default: 5:10.}

\item{alp}{Level of sinificance.}

\item{ws}{Graphic window size, Default: c(7,4).}

\item{sep}{Logical: plot in separate facets? Default: FALSE.}
}
\value{
None.
}
\description{
Plots and Tables of the Distribution of Wilcoxon Signed-Rank Test Statistic.
}
\examples{
signrank.dist(nv=5:15)
signrank.dist(nv=c(5,7,10,20))
}
