% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{signrank.table}
\alias{signrank.table}
\title{Quantiles of Wilcoxon Signed-Rank Test Statistic}
\usage{
signrank.table(nv = 5:10, pp)
}
\arguments{
\item{nv}{Number of signed data, Default: 5:10}

\item{pp}{Probability vector, Default: c(0.95, 0.975, 0.99, 0.995))}
}
\value{
None.
}
\description{
Table of the Quantiles of Wilcoxon Signed-Rank Test Statistic
}
\examples{
signrank.table(nv=5:15)
pp <- c(0.005, 0.01, 0.025, 0.05, 0.95, 0.975, 0.99, 0.995)
signrank.table(nv=5:15, pp=pp)
signrank.table(nv=c(5,7,10,20))
}
