% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{signrank.tsp}
\alias{signrank.tsp}
\title{Wilcoxon Signed-Rank Test}
\usage{
signrank.tsp(x, y, mu0 = 0, alt = "two", dig = 4, ws = "n")
}
\arguments{
\item{x}{Data of the first group.}

\item{y}{Data of the second group.}

\item{mu0}{Mean difference under H0, Default: 0.}

\item{alt}{Type of alternative hypothesis, Default: 'two'.}

\item{dig}{Number of decimal places, Default: 4.}

\item{plot}{Graphic window size, Default: "n".}
}
\value{
None.
}
\description{
Wilcoxon Signed-Rank Test with a Plot
}
\examples{
x <- c(38, 26, 34, 5, 68, 30, 35, 19, 33, 69)
y <- c(28, 21, 31, 11, 59, 28, 28, 23, 32, 38)
d <- x-y; shapiro.test(d)
signrank.tsp(x, y, alt="gr", ws=c(7,4))
signrank.tsp(y, x, alt="le", ws=c(7,4))
}
