% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{signtest.p}
\alias{signtest.p}
\title{Sign Test (Binomial Test)}
\usage{
signtest.p(x, mu0 = 0, alt = "two", dig = 4, ws = c(7, 4))
}
\arguments{
\item{x}{Data vector.}

\item{mu0}{Mean value under H0, Default: 0.}

\item{alt}{Type of alternative hypothesis, Default: 'two'.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}
}
\value{
None.
}
\description{
Sign Test (Binomial Test) with a Plot
}
\examples{
(x <- c(1,2,5,7, rep(8,7), rep(9,5), rep(10,4)))
signtest.p(x=x, mu0=7, alt="gr")
signtest.p(x=x, mu0=7)
signtest.p(x=x, mu0=9, alt="le")
}
