% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{snorm.cdf}
\alias{snorm.cdf}
\title{Plot Standard Normal Cumulative Probability}
\usage{
snorm.cdf(zp = -2:2, prt = TRUE, dig = 4, ws = c(7, 4), ...)
}
\arguments{
\item{zp}{Values of z for annotation, Default: -2:2.}

\item{prt}{Logical: print the output? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Grapic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\description{
Plot Standard Normal Cumulative Probability P(Z<z)
}
\examples{
snorm.cdf()

zp <- seq(-2, 2, by=0.5)
snorm.cdf(zp)
}
