% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{snorm.prob}
\alias{snorm.prob}
\title{Interval Probability of the Standard Normal Distribution}
\usage{
snorm.prob(lb = -(1:4), ub = -lb, prt = TRUE, dig = 4, ws = c(7, 4), ...)
}
\arguments{
\item{lb}{Lower bound of z values, Default: -(1:4).}

\item{ub}{Upper bound of z values, Default: -lb.}

\item{prt}{Logical: print the output? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Grapic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
Interval probabilities.
}
\description{
Interval Probability of the Standard Normal Distribution.
}
\examples{
zp <- 1:3
snorm.prob(zp)

snorm.prob(-4:0*1.5, 0:4*1.5)
}
