% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{snorm.quant1}
\alias{snorm.quant1}
\title{Quantile Plot of the Standard Normal Distribution}
\usage{
snorm.quant1(pv, pv2 = pv, dig = 4, main)
}
\arguments{
\item{pv}{Vector of probability values}

\item{pv2}{Vector of specific probability values}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{main}{Graph title, Default: 'Quantiles of the Standard Normal Distribution'}
}
\value{
None.
}
\description{
Quantile Plot of the Standard Normal Distribution (revised)
}
\examples{
pv <- c(0.005, 0.01, 0.025, 0.05, 0.5, 0.95, 0.975, 0.99, 0.995)
# Step type
snorm.quant1(pv)

}
