% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{snorm.quant2}
\alias{snorm.quant2}
\title{Quantile Plot of the Standard Normal Distribution}
\usage{
snorm.quant2(pv, dig = 4, main, ws = c(7, 6))
}
\arguments{
\item{pv}{Vector of probability values}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{main}{Graph title, Default: 'Quantiles of the Standard Normal Distribution'}

\item{pv2}{Vector of specific probability values}
}
\value{
None.
}
\description{
Quantile Plot of the Standard Normal Distribution (CDF version)
}
\examples{
pv <- c(0.005, 0.01, 0.025, 0.05, 0.5, 0.95, 0.975, 0.99, 0.995)
# CDF type
snorm.quant2(pv)

}
