% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{sreg.lse}
\alias{sreg.lse}
\title{Simple Regression: Least Square Estimation}
\usage{
sreg.lse(x, y, data, type = "lse", alp = 0.05, dig = 4, dp, ws = "n", ...)
}
\arguments{
\item{x}{Independent (explanatory) variable (or formula).}

\item{y}{Dependent (response) variable.}

\item{data}{Data frame applied to x and y (or formula).}

\item{type}{One of ("lse", "test", "aov", "all").}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of decimal places, Default: 4.}

\item{dp}{Selection number of diagnostic plots.}

\item{ws}{Graphic window size, Default: "n".}

\item{...}{Other graphic parameters.}
}
\value{
Least square estimates, ANOVA table, or test statistics
}
\description{
Calculate the least square estimates, test on the significance, or test on the coefficients for a simple regression.
}
\examples{
sreg.lse(Petal.Length ~ Sepal.Length, data=iris)
sreg.lse(1, 3, data=iris)
sreg.lse(1, 3, data=iris, type="aov")
sreg.lse(1, 3, data=iris, type="test")
sreg.lse(1, 3, data=iris, type=c("lse","aov"), dp=c(1:3,5))

x <- c(1095, 1110, 1086, 1074, 1098, 1105, 1163, 1124, 1088, 1064)
y <- c(49, 52, 48, 49, 50, 51, 50, 51, 49, 48)
(y2 <- y*x/1000)
sreg.lse(x, y2)
sreg.lse(y2 ~ x)
sreg.lse(y2 ~ x, type="aov")
sreg.lse(y2 ~ x, type="test")
sreg.lse(y2 ~ x, type="all", dp=c(1:3,5))
}
