% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{sreg.pred}
\alias{sreg.pred}
\title{Prediction and Confidence Interval}
\usage{
sreg.pred(x, y, data, x0, alp = 0.05, dig = 4, ws = "n", ...)
}
\arguments{
\item{x}{Independent variable (or formula or lm() object).}

\item{y}{Dependent (response) variable.}

\item{data}{Data frame applied to x and y (or formula).}

\item{x0}{Conditional value of the independent variable.}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}

\item{...}{Other graphic parameters.}
}
\value{
CI and PI.
}
\description{
Prediction and Confidence Interval in a Simple Regression.
}
\examples{
sreg.pred(iris[3], iris[4], x0=2)
sreg.pred(iris[3], iris[4], x0=2, plot=T)

x <- c(1095, 1110, 1086, 1074, 1098, 1105, 1163, 1124, 1088, 1064)
y <- c(49, 52, 48, 49, 50, 51, 50, 51, 49, 48)
(y2 <- y*x/1000)
sreg.pred(x, y2, x0=1200, plot=T)

}
