% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{strat.hist}
\alias{strat.hist}
\title{Stratified Histograms}
\usage{
strat.hist(ng = 3, n = 200, mu = 2 + 4 * 1:ng, sig = 1, spec, ws, mfd, ...)
}
\arguments{
\item{ng}{Number of subgroups, Default: 3.}

\item{n}{Number of data for each subgroup, Default: 200.}

\item{mu}{Means of subgroups, Default: 2+4*1:ng.}

\item{sig}{Standard deviations of subgroups, Default: 1.}

\item{spec}{Specification limits, Default: c(min(m)-2, max(m)+2).}

\item{ws}{Graphic window size vector.}

\item{mfd}{Multi-frame dimension vector}

\item{...}{Other graphic parameters.}
}
\value{
None
}
\description{
To create Stratified Histograms from a mixed data set.
}
\examples{
strat.hist()
strat.hist(ng=4, mu=c(6,10,14,18), ws=c(9,6), spec=c(4,20))
strat.hist(ng=5, n=400, ws=c(9,6), prob=T)
}
\keyword{Ch2}
\keyword{Descriptive}
\keyword{Statistics.}
