% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{tdist.prob}
\alias{tdist.prob}
\title{Interval Probabilities of the T-Distribution}
\usage{
tdist.prob(nu, lb = -(1:4), ub = -lb, prt = TRUE, dig = 4, ws = c(7, 4), ...)
}
\arguments{
\item{nu}{Degrees of freedom.}

\item{lb}{Lower bound of x values, Default: -(1:4).}

\item{ub}{Upper bound of x values, Default: -lb.}

\item{prt}{Logical: print the output? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Grapic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
Interval Probabilities.
}
\description{
Interval Probability of the T-Distribution.
}
\examples{
xp <- 1:3
tdist.prob(nu=10, lb=-xp)

tdist.prob(nu=15, -4:0*1.5, 0:4*0.5)
}
