% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{tdist.sim}
\alias{tdist.sim}
\title{Simulation for the t-distribution}
\usage{
tdist.sim(
  ns,
  mu = 0,
  sig = 1,
  N = 1000,
  prt = TRUE,
  ws = c(10, 4),
  dig = 4,
  ...
)
}
\arguments{
\item{ns}{Sample size.}

\item{mu}{Expected value, Default: 0.}

\item{sig}{Standard deviation, Default: 1.}

\item{N}{Number of iterations, Default: 1000.}

\item{prt}{Logical: print statistics? Default: TRUE.}

\item{ws}{Graphic window size, Default: c(10,4).}

\item{dig}{Number of decimal places, Default: 4.}

\item{...}{Other graphic parameters.}
}
\value{
Sample means and standard deviations.
}
\description{
Simulation for the t-distribution.
}
\examples{
tdist.sim(ns=10, mu=100, sig=10)

tdist.sim(ns=5, N=10000, lwd=3)

tdist.sim(ns=5, breaks=100)

tdist.sim(ns=5, mu=100, sig=10, N=10000, breaks=200, seed=157)

}
