% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{tnorm.comp}
\alias{tnorm.comp}
\title{Compare T-distribution with the Standard Normal}
\usage{
tnorm.comp(nu, ws = c(9, 4), ...)
}
\arguments{
\item{nu}{Degree of freedom for the T-distribution.}

\item{ws}{Grapic window size, Default: c(10,4).}

\item{...}{Other graphic parameters.}
}
\value{
None.
}
\description{
Compare T-distribution with the Standard Normal
}
\examples{
nu <- c(1, 5, 10, 30)
tnorm.comp(nu)
tnorm.comp(10:100)
}
