% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{ttest.plot}
\alias{ttest.plot}
\title{Plot the PDF of Test Statistic with the T-distribution}
\usage{
ttest.plot(
  t0,
  df,
  alt = "two",
  dig = 4,
  mt,
  prng = c(-4, 4),
  xlab,
  prt = FALSE
)
}
\arguments{
\item{t0}{T-test statistic for the population mean}

\item{df}{Degree of freedom}

\item{alt}{Type of the alternative hypothesis, Default: 'two'}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{mt}{Main title of the plot}

\item{prng}{Range of x-axis, Default: c(-4, 4)}

\item{xlab}{Label of x-axis, Default: 'Test Statistic (T0)'}

\item{prt}{Logical value for printing the p-value, Default: FALSE}
}
\value{
None.
}
\description{
Plot the PDF of Test Statistic with the T-distribution
}
\examples{
win.graph(7, 5)
ttest.plot(1.96, df=24)

set.seed(3691)
x <- rnorm(20, 199, 2)
y <- rnorm(25, 200, 4)
(res <- t.test(x,y))
ttest.plot(res$stat, res$para)
}
