% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{unstable.hist}
\alias{unstable.hist}
\title{Unstable Histograms}
\usage{
unstable.hist(
  N = 200,
  mu = c(10, 6),
  sig = c(1, 0.5),
  miss = c(8, 9),
  cut = 9,
  ws,
  mfd,
  ...
)
}
\arguments{
\item{N}{Number of random variates, Default: 200.}

\item{mu}{Means of two normal distributions, Default: c(10,6).}

\item{sig}{Standard deviations of two distributions, Default: c(1,0.5).}

\item{miss}{Limits of the missing values, Default: c(8,9).}

\item{cut}{Cut-off value, Default: 9.}

\item{ws}{Graphic window size vector.}

\item{mfd}{Multi-frame dimension vector.}

\item{...}{Other graphic parameters.}
}
\value{
None
}
\description{
To create four types of unstable histograms.
}
\examples{
unstable.hist()
unstable.hist(mu=c(8,15), sig=c(1.5,0.6), miss=c(9,10), cut=10, col=2:5)
unstable.hist(col=2:5, ws=c(12,4), mfd=c(1,4))
}
\keyword{Ch2}
\keyword{Descriptive}
\keyword{Statistics}
