% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{urnsamp.sum}
\alias{urnsamp.sum}
\title{PDF of the Sum of n Numbers from a Finite Population}
\usage{
urnsamp.sum(x, n, repl = FALSE, prt = FALSE, dig = 4, ws = c(7, 5), mt, cex)
}
\arguments{
\item{x}{Numeric vector representing the population.}

\item{n}{Sample size.}

\item{repl}{Logical: sampling with replacement? Default: FALSE.}

\item{prt}{Logical: print the distribution? Default: FALSE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,5).}

\item{mt}{Main title of the plot.}

\item{cex}{Text size in the plot.}
}
\value{
List of frequency and probability distribution.
}
\description{
Probability Distribution of the Sum of n Numbers Drawn from a Finite Population.
}
\examples{
# Sum of three samples out of 1:10
urnsamp.sum(1:10, 3)

# Rolling four dice
mytitle <- "PDF of the sum of Four Dice"
urnsamp.sum(1:6, 4, repl=TRUE, mt=mytitle)

# Tossing 20 coins, sum of heads
mytitle <- "PDF of the sum of Heads in 20 Coin Tosses"
out <- urnsamp.sum(0:1, 20, repl=TRUE, mt=mytitle)
sum(out$freq); 2^20

}
