% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{var1.test}
\alias{var1.test}
\title{Hypothesis Test for a Variance}
\usage{
var1.test(x, n, var0, alt = "two", alp = 0.05, dig = 4, ws = c(7, 4), ...)
}
\arguments{
\item{x}{Sample variance or sample data}

\item{n}{Sample size (necessary only if no data are given).}

\item{var0}{Population variance under H0 (if missing, just find the CI).}

\item{alt}{Type of the alternative hypothesis ("gr", "le", "two"), Default: "two"}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of digits below the decimal point, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
Summary Statistics.
}
\description{
Hypothesis Test for a Population Variance
}
\examples{
var1.test(x=1.24, n=25)
var1.test(x=1.24, n=25, alt="gr")
var1.test(x=1.24, n=25, var0=0.8, alt="gr")

x <- c(20.0, 21.5, 20.9, 19.8, 22.5, 20.3, 23.6, 18.0, 23.3, 17.8)
var1.test(x=x, alt="two")
var1.test(x=x, var0=2, alt="two")

set.seed(1234)
x <- rnorm(40, 10, 1.5)
var1.test(x, var0=2, alt="two")

}
