% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{var2.test}
\alias{var2.test}
\title{Inference on the Ratio of Two Population Variances}
\usage{
var2.test(x, y, n1, n2, alp = 0.05, alt = "two", dig = 4, ws = c(7, 4), ...)
}
\arguments{
\item{x}{First sample data or sample variance.}

\item{y}{Second sample data or sample variance.}

\item{n1}{First sample size (without sample data).}

\item{n2}{Second sample size (without sample data).}

\item{alp}{Level of significance, Default: 0.05}

\item{alt}{Type of the alternative hypothesis, Default: 'two'}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,4).}

\item{...}{Other graphic parameters.}
}
\value{
list(stat=F0, df=c(df1,df2), cval=cv, pval=pv)
}
\description{
Inference on the Ratio of Two Independent Normal Population Variances.
}
\examples{
var2.test(x=3.5^2, y=5.8^2, n1=25, n2=34)

set.seed(1234)
x <- rnorm(20, 199, 4)
y <- rnorm(25, 200, 2)
var2.test(x, y)
var2.test(x, y, alt="gr")
var2.test(var(x), var(y), length(x), length(y), alt="gr")
}
