% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{venn.count}
\alias{venn.count}
\title{Venn Diagram of Three Events}
\usage{
venn.count(A, B, C, N, col = 2:4, alp = 0.2, ws = c(7, 4))
}
\arguments{
\item{A}{Event A (data frame or vector).}

\item{B}{Event B (data frame or vector).}

\item{C}{Event C (data frame or vector).}

\item{N}{Total number of elements.}

\item{col}{Event colors, Default: c(2,3,4).}

\item{alp}{Transparency of colors, Default: 0.2.}

\item{ws}{The size of the graphic window, Default: c(7,4).}
}
\value{
The number of elements for c(A, B, C, AB, AC, BC, ABC, S, S/ABC).
}
\description{
Venn Diagram of Three Events with the Number of Elements
}
\examples{
S <- rolldie2(2); N <- nrow(S)
A <- subset(S, ((X1+X2) \%\% 2)==0); element(A)
B <- subset(S, (X1+X2) >= 8); element(B)
C <- subset(S, abs(X1-X2) <= 1); element(C)
venn.count(A, B, C, N)
}
