% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{venn.count2}
\alias{venn.count2}
\title{Count the intersections of up to four events}
\usage{
venn.count2(S, EV, ws = c(7, 4))
}
\arguments{
\item{S}{Data frame of a sample space.}

\item{EV}{List of events to draw and count.}

\item{ws}{Size of graphic window, Default: c(7,4).}
}
\value{
The vector of counts for the intersections.
}
\description{
Draw a Venn diagram and count the intersections of events.
}
\examples{
S = rolldie2(2)
S = S[order(S[[1]], S[[2]]), ]
A = subset(S, ((X1+X2) \%\% 2)==0); element(A)
B = subset(S, (X1+X2)>=8); element(B)
C = subset(S, abs(X1-X2)<=1); element(C)
D = subset(S, (X1*X2)<12); element(D)
venn.count2(S, EV=list(A,B,C,D))
}
