% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.cdf}
\alias{cont.cdf}
\title{CDF of a Continuous Random Variable}
\usage{
cont.cdf(fun, lb, ub, xp, prt = FALSE, dig = 4, ws = c(7, 5), xlim, ...)
}
\arguments{
\item{lb}{Lower bound of the PDF.}

\item{ub}{Upper bound of the PDF.}

\item{xp}{Specific values of X for displaying the probability.}

\item{prt}{Logical: print F(xp)? Default: FALSE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: c(7,5).}

\item{xlim}{Lower and Upper limit of x-axis.}

\item{...}{Other graphic parameters.}

\item{fx}{Probability density function f(x) of X.}
}
\value{
list(CDF=CDF, mean=EX, sd=DX).
}
\description{
Cumulative Distribution Function of a Continuous Random Variable.
}
\examples{
# f(x) = 2*exp(-2*x), for (x>0)
cont.cdf("2*exp(-2*x)", lb=0, xp=c((1:5)*0.2, 2), xlim=c(-1,3))
cont.cdf("2*exp(-2*x)", lb=0, xp=c((1:5)*0.2, 2), prt=TRUE, ws="n")
cont.cdf("2*exp(-2*x)", lb=0, xp=c((1:5)*0.2, 2), prt=TRUE)

# f(x) = 3*x^2, for (0<x<1)
cont.cdf("3*x^2", lb=0, ub=1, xp=(2:5)*0.2, prt=TRUE)

# f(x) = dnorm(x,10,2), for (-Inf<x<Inf)
cont.cdf("dnorm(x,10,2)", xp=3:7*2, prt=TRUE)

}
