% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch7-fn.R
\name{cont.dist}
\alias{cont.dist}
\alias{cont.mpdf}
\title{Expectations, Transformation, and PDF Plots for Continuous Distributions}
\usage{
cont.dist(dist, para1, para2, Tr, prt = "var", dig = 4, ws = "n", ...)

cont.mpdf(dist, para1, para2, xlim, ymax, xp1, xp2, np = 100, ws, ...)
}
\arguments{
\item{dist}{Name of continuous probability distribution (one of the follows)
("exp", "gamma", "weibull", "beta", "norm", "t", "chisq", "f").}

\item{para1}{First parameter vector of the distribution.}

\item{para2}{Second parameter vector of the distribution (optional).}

\item{Tr}{Transformation function in text string.}

\item{prt}{Print option, one of c("exp", "var", "n"), Default: "var".}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graph window size, Default: c(9,5).}

\item{...}{Other graphic parameters.}

\item{xlim}{Limit of x-axis for the plot.}

\item{ymax}{Upper limit of y-axis for the plot.}

\item{xp1}{Specific x values for annotations in the PDF plot.}

\item{xp2}{Specific x values for annotations in the CDF plot.}

\item{np}{Number of plot points, Default: 100.}
}
\value{
list(fun, Ex, Vx, Tfun, Ey, Vy)

None.
}
\description{
Expectation, Variance, Transformation, and PDF Plots for Continuous Distributions.

PDF and CDF Plots for Continuous Distributions.
}
\examples{
cont.dist("unif", 0, 1, Tr="x^2", prt="var", ws=c(9,4))

cont.dist("exp", 5, Tr="2*x+4", prt="var", ws=c(9,4))

cont.dist("gamma", 2, 5, Tr="2*(x-5)", prt="var", ws=c(9,4))

cont.dist("weib", 2, 5, Tr="x^2", prt="var", ws=c(9,4))

cont.dist("beta", 2, 5, Tr="2+4*x", prt="var", ws=c(9,4))

lamb <- 1:5
cont.mpdf("exp", para1=lamb, xlim=3, ymax=5)

alp <- c(0.5, 1, 2, 3); rate <- 1
cont.mpdf("gamma", alp, rate, xlim=8, ymax=1.2)

th <- 1; alp <- c(0.5, 1, 2, 3)
cont.mpdf("weibull", alp, th, 5, 1.2)

alp <- 10:100/10; th <- 1
cont.mpdf("gamma", alp, th, 20)
cont.mpdf("weibull", alp, th, 5)
}
