% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{disc.exp}
\alias{disc.exp}
\title{Expected Values of a Discrete Random Variable}
\usage{
disc.exp(xf, xv, Tr, prt = "exp", dig = 4, ws = "n", ...)
}
\arguments{
\item{xf}{Probability (frequency) distribution or random variable.}

\item{xv}{Values of X (required if xf is a function or vector).}

\item{Tr}{Transformation formula in string.}

\item{prt}{Print option: one of c("exp", "var"), Default: "exp".}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}

\item{...}{Other graphic parameters.}
}
\value{
list(EX=expected value, VX=variance)
}
\description{
Expected Values of a Discrete Random Variable
}
\examples{
S <- tosscoin2(10)
X <- apply(S,1,\(x) sum(x=="H"))
x <- table(X)
disc.exp(x, ws=c(7,4))

x <- (0:20); y <- x^2
disc.exp(dbinom(x, 20, 0.5), y, prt="var", ws=c(7,4), cex=0.6)
}
