% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{fdist.sim}
\alias{fdist.sim}
\title{Simulation for the F-distribution}
\usage{
fdist.sim(
  sig1 = 1,
  sig2 = 1,
  n1,
  n2,
  N = 1000,
  prt = TRUE,
  ws = c(10, 4),
  dig = 4,
  ...
)
}
\arguments{
\item{sig1}{Standard deviation of the first population.}

\item{sig2}{Standard deviation of the second population.}

\item{n1}{Sample size of the first population.}

\item{n2}{Sample size of the second population.}

\item{N}{Number of iterations, Default: 1000.}

\item{prt}{Logical: print statistics? Default: TRUE.}

\item{ws}{Graphic window size, Default: c(10,4).}

\item{dig}{Number of decimal places, Default: 4.}

\item{...}{Other graphic parameters.}
}
\value{
Sample means and standard deviations.
}
\description{
Simulation for the F-distribution.
}
\examples{
# Stable
fdist.sim(sig1=2, sig2=7, n1=15, n2=20)
# Very unstable
fdist.sim(n1=5, n2=4, xlim=c(0,20))
fdist.sim(n1=5, n2=4, N=10000, breaks=100, xlim=c(0,20))

}
