% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{mreg.lse}
\alias{mreg.lse}
\title{Calculate Least Square Estimates: Multiple Regression}
\usage{
mreg.lse(form, data, type = "lse", alp = 0.05, dig = 4, ws = "n")
}
\arguments{
\item{form}{Formula of the regression model (ex: y ~ x1 + x2)}

\item{data}{Data frame applied to the formula.}

\item{type}{Type of analysis, one of ("lse", "test", "aov", "all").}

\item{alp}{Level of significance, Default: 0.05.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Graphic window size, Default: "n".}

\item{test}{Logical value for testing on the coefficients, Default: FALSE.}
}
\value{
Matrices for calculating the LSE.
}
\description{
Calculate Least Square Estimates for a Multiple Regression.
}
\examples{
attach(mtcars)
form <- mpg ~ hp + drat + wt
mreg.lse(form)
mreg.lse(form, pre=NA, aov=TRUE)
mreg.lse(form, pre=NA, test=TRUE)

data(exa14_10)
Score <- exa14_10[[1]]; Study <- exa14_10[[2]]; Reading <- exa14_10[[3]]
form2 <- Score ~ Study + Reading
mreg.lse(form2)
mreg.lse(form2, pre=NA, aov=TRUE)
mreg.lse(form2, pre=NA, test=TRUE)
}
