% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{norm.trans}
\alias{norm.trans}
\title{Standardization of the Normal Distribution}
\usage{
norm.trans(mu, sig, a, b, prt = TRUE, dig = 4, ws = c(9, 4), ...)
}
\arguments{
\item{mu}{Mean of the normal distribution.}

\item{sig}{Standard deviation of the normal distribution.}

\item{a}{Lower limit of X for P(a<X<b).}

\item{b}{Upper limit of X for P(a<X<b).}

\item{prt}{Logical: print the output? Default: TRUE.}

\item{dig}{Number of decimal places, Default: 4.}

\item{ws}{Grapic window size, Default: c(9,4).}

\item{...}{Other graphic parameters.}
}
\value{
None.
}
\description{
Check Probability Conservation in Standardizing the Normal Distribution
}
\examples{
norm.trans(2, 2, -1, 4)

norm.trans(100, 4, 95, 108)
}
